@@ -1,93 +1,241 @@
 [package]
 name = "iced"
-version = "0.1.1"
-authors = ["Héctor Ramón Jiménez <hector0193@gmail.com>"]
-edition = "2018"
 description = "A cross-platform GUI library inspired by Elm"
-license = "MIT"
-repository = "https://github.com/hecrj/iced"
-documentation = "https://docs.rs/iced"
-readme = "README.md"
-keywords = ["gui", "ui", "graphics", "interface", "widgets"]
-categories = ["gui"]
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+categories.workspace = true
+keywords.workspace = true
+rust-version.workspace = true
 
-[features]
-default = ["wgpu", "default_system_font"]
-# Enables the `iced_wgpu` renderer
-wgpu = ["iced_wgpu"]
-# Enables the `Image` widget
-image = ["iced_wgpu/image"]
-# Enables the `Svg` widget
-svg = ["iced_wgpu/svg"]
-# Enables the `Canvas` widget
-canvas = ["iced_wgpu/canvas"]
-# Enables using system fonts.
-default_system_font = ["iced_wgpu/default_system_font"]
-# Enables the `iced_glow` renderer. Overrides `iced_wgpu`
-glow = ["iced_glow", "iced_glutin"]
-# Enables the `Canvas` widget for `iced_glow`
-glow_canvas = ["iced_glow/canvas"]
-# Enables using system fonts for `iced_glow`.
-glow_default_system_font = ["iced_glow/default_system_font"]
-# Enables a debug view in native platforms (press F12)
-debug = ["iced_winit/debug"]
-# Enables `tokio` as the `executor::Default` on native platforms
-tokio = ["iced_futures/tokio"]
-# Enables `async-std` as the `executor::Default` on native platforms
-async-std = ["iced_futures/async-std"]
-# Enables advanced color conversion via `palette`
-palette = ["iced_core/palette"]
+[lints]
+workspace = true
+
+[package.metadata.docs.rs]
+rustdoc-args = ["--cfg", "docsrs"]
+all-features = true
 
 [badges]
 maintenance = { status = "actively-developed" }
 
-[workspace]
-members = [
-    "core",
-    "futures",
-    "graphics",
-    "glow",
-    "glutin",
-    "native",
-    "style",
-    "web",
-    "wgpu",
-    "winit",
-    "examples/bezier_tool",
-    "examples/clock",
-    "examples/color_palette",
-    "examples/counter",
-    "examples/custom_widget",
-    "examples/download_progress",
-    "examples/events",
-    "examples/game_of_life",
-    "examples/geometry",
-    "examples/integration",
-    "examples/pane_grid",
-    "examples/pick_list",
-    "examples/pokedex",
-    "examples/progress_bar",
-    "examples/solar_system",
-    "examples/stopwatch",
-    "examples/styling",
-    "examples/svg",
-    "examples/todos",
-    "examples/tour",
-]
+[features]
+default = ["wgpu", "tiny-skia", "web-colors", "auto-detect-theme", "thread-pool"]
+# Enables the `wgpu` GPU-accelerated renderer backend
+wgpu = ["iced_renderer/wgpu", "iced_widget/wgpu"]
+# Enables the `tiny-skia` software renderer backend
+tiny-skia = ["iced_renderer/tiny-skia"]
+# Enables the `image` widget
+image = ["image-without-codecs", "image/default"]
+# Enables the `image` widget, without any built-in codecs of the `image` crate
+image-without-codecs = ["iced_widget/image", "dep:image"]
+# Enables the `svg` widget
+svg = ["iced_widget/svg"]
+# Enables the `canvas` widget
+canvas = ["iced_widget/canvas"]
+# Enables the `qr_code` widget
+qr_code = ["iced_widget/qr_code"]
+# Enables the `markdown` widget
+markdown = ["iced_widget/markdown"]
+# Enables lazy widgets
+lazy = ["iced_widget/lazy"]
+# Enables a debug view in native platforms (press F12)
+debug = ["iced_winit/debug", "iced_devtools"]
+# Enables time-travel debugging (very experimental!)
+time-travel = ["debug", "iced_devtools/time-travel"]
+# Enables the `thread-pool` futures executor as the `executor::Default` on native platforms
+thread-pool = ["iced_futures/thread-pool"]
+# Enables `tokio` as the `executor::Default` on native platforms
+tokio = ["iced_futures/tokio"]
+# Enables `smol` as the `executor::Default` on native platforms
+smol = ["iced_futures/smol"]
+# Enables querying system information
+system = ["iced_winit/system"]
+# Enables broken "sRGB linear" blending to reproduce color management of the Web
+web-colors = ["iced_renderer/web-colors"]
+# Enables pixel snapping for crisp edges by default (can cause jitter!)
+crisp = ["iced_core/crisp", "iced_widget/crisp"]
+# Enables the WebGL backend
+webgl = ["iced_renderer/webgl"]
+# Enables syntax highligthing
+highlighter = ["iced_highlighter", "iced_widget/highlighter"]
+# Enables the advanced module
+advanced = ["iced_core/advanced", "iced_widget/advanced"]
+# Embeds Fira Sans into the final application; useful for testing and Wasm builds
+fira-sans = ["iced_renderer/fira-sans"]
+# Auto-detects light/dark mode for the built-in theme
+auto-detect-theme = ["iced_core/auto-detect-theme"]
+# Enables strict assertions for debugging purposes at the expense of performance
+strict-assertions = ["iced_renderer/strict-assertions"]
+# Redraws on every runtime event, and not only when a widget requests it
+unconditional-rendering = ["iced_winit/unconditional-rendering"]
+# Enables support for the `sipper` library
+sipper = ["iced_runtime/sipper"]
 
 [dependencies]
-iced_core = { version = "0.2", path = "core" }
-iced_futures = { version = "0.1", path = "futures" }
+iced_debug.workspace = true
+iced_core.workspace = true
+iced_futures.workspace = true
+iced_renderer.workspace = true
+iced_runtime.workspace = true
+iced_widget.workspace = true
+iced_winit.features = ["program"]
+iced_winit.workspace = true
 
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-iced_winit = { version = "0.1", path = "winit" }
-iced_glutin = { version = "0.1", path = "glutin", optional = true }
-iced_wgpu = { version = "0.2", path = "wgpu", optional = true }
-iced_glow = { version = "0.1", path = "glow", optional = true}
+iced_devtools.workspace = true
+iced_devtools.optional = true
 
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-iced_web = { version = "0.2", path = "web" }
+iced_highlighter.workspace = true
+iced_highlighter.optional = true
 
-[package.metadata.docs.rs]
-rustdoc-args = ["--cfg", "docsrs"]
-features = ["image", "svg", "canvas"]
+thiserror.workspace = true
+
+image.workspace = true
+image.optional = true
+
+[dev-dependencies]
+criterion = "0.5"
+iced_wgpu.workspace = true
+
+[[bench]]
+name = "wgpu"
+harness = false
+required-features = ["canvas"]
+
+[profile.release-opt]
+inherits = "release"
+codegen-units = 1
+debug = false
+lto = true
+incremental = false
+opt-level = 3
+overflow-checks = false
+strip = "debuginfo"
+
+[workspace]
+members = [
+    "beacon",
+    "core",
+    "debug",
+    "devtools",
+    "futures",
+    "graphics",
+    "highlighter",
+    "program",
+    "renderer",
+    "runtime",
+    "test",
+    "tiny_skia",
+    "wgpu",
+    "widget",
+    "winit",
+    "examples/*",
+]
+
+[workspace.package]
+version = "0.14.0-dev"
+authors = ["Héctor Ramón Jiménez <hector@hecrj.dev>"]
+edition = "2024"
+license = "MIT"
+repository = "https://github.com/iced-rs/iced"
+homepage = "https://iced.rs"
+categories = ["gui"]
+keywords = ["gui", "ui", "graphics", "interface", "widgets"]
+rust-version = "1.85"
+
+[workspace.dependencies]
+iced = { version = "0.14.0-dev", path = "." }
+iced_beacon = { version = "0.14.0-dev", path = "beacon" }
+iced_core = { version = "0.14.0-dev", path = "core" }
+iced_debug = { version = "0.14.0-dev", path = "debug" }
+iced_devtools = { version = "0.14.0-dev", path = "devtools" }
+iced_futures = { version = "0.14.0-dev", path = "futures" }
+iced_graphics = { version = "0.14.0-dev", path = "graphics" }
+iced_highlighter = { version = "0.14.0-dev", path = "highlighter" }
+iced_program = { version = "0.14.0-dev", path = "program" }
+iced_renderer = { version = "0.14.0-dev", path = "renderer" }
+iced_runtime = { version = "0.14.0-dev", path = "runtime" }
+iced_test = { version = "0.14.0-dev", path = "test" }
+iced_tiny_skia = { version = "0.14.0-dev", path = "tiny_skia" }
+iced_wgpu = { version = "0.14.0-dev", path = "wgpu" }
+iced_widget = { version = "0.14.0-dev", path = "widget" }
+iced_winit = { version = "0.14.0-dev", path = "winit" }
+
+bincode = "1.3"
+bitflags = "2.0"
+bytemuck = { version = "1.0", features = ["derive"] }
+bytes = "1.6"
+cosmic-text = "0.14"
+dark-light = "2.0"
+futures = { version = "0.3", default-features = false }
+glam = "0.25"
+cryoglyph = { git = "https://github.com/iced-rs/cryoglyph.git", rev = "a456d1c17bbcf33afcca41d9e5e299f9f1193819" }
+guillotiere = "0.6"
+half = "2.2"
+image = { version = "0.25", default-features = false }
+kamadak-exif = "0.5"
+kurbo = "0.10"
+lilt = "0.8"
+log = "0.4"
+lyon = "1.0"
+lyon_path = "1.0"
+num-traits = "0.2"
+ouroboros = "0.18"
+png = "0.17"
+pulldown-cmark = "0.12"
+qrcode = { version = "0.13", default-features = false }
+raw-window-handle = "0.6"
+resvg = "0.42"
+rustc-hash = "2.0"
+serde = "1.0"
+semver = "1.0"
+sha2 = "0.10"
+sipper = "0.1"
+smol = "2"
+smol_str = "0.2"
+softbuffer = "0.4"
+syntect = "5.1"
+sysinfo = "0.33"
+thiserror = "1.0"
+tiny-skia = "0.11"
+tokio = "1.0"
+tracing = "0.1"
+unicode-segmentation = "1.0"
+url = "2.5"
+wasm-bindgen-futures = "0.4"
+wasmtimer = "0.4.1"
+web-sys = "0.3.69"
+web-time = "1.1"
+wgpu = "24.0"
+window_clipboard = "0.4.1"
+winit = { git = "https://github.com/iced-rs/winit.git", rev = "11414b6aa45699f038114e61b4ddf5102b2d3b4b" }
+
+[workspace.lints.rust]
+rust_2018_idioms = { level = "deny", priority = -1 }
+missing_debug_implementations = "deny"
+missing_docs = "deny"
+unsafe_code = "deny"
+unused_results = "deny"
+
+[workspace.lints.clippy]
+type-complexity = "allow"
+map-entry = "allow"
+large-enum-variant = "allow"
+result_large_err = "allow"
+semicolon_if_nothing_returned = "deny"
+trivially-copy-pass-by-ref = "deny"
+default_trait_access = "deny"
+match-wildcard-for-single-variants = "deny"
+redundant-closure-for-method-calls = "deny"
+filter_map_next = "deny"
+manual_let_else = "deny"
+unused_async = "deny"
+from_over_into = "deny"
+needless_borrow = "deny"
+new_without_default = "deny"
+useless_conversion = "deny"
+
+[workspace.lints.rustdoc]
+broken_intra_doc_links = "forbid"
