@@ -2,9 +2,11 @@
 name = "color_palette"
 version = "0.1.0"
 authors = ["Clark Moody <clark@clarkmoody.com>"]
-edition = "2018"
+edition = "2024"
 publish = false
 
 [dependencies]
-iced = { path = "../..", features = ["canvas", "palette"] }
-palette = "0.5.0"
+iced.workspace = true
+iced.features = ["canvas"]
+
+palette = "0.7"
