@@ -1,12 +1,14 @@
 [package]
 name = "download_progress"
 version = "0.1.0"
-authors = ["Songtronix <contact@songtronix.com>"]
-edition = "2018"
+authors = ["Songtronix <contact@songtronix.com>", "Folyd <lyshuhow@gmail.com>"]
+edition = "2024"
 publish = false
 
 [dependencies]
-iced = { path = "../..", features = ["tokio"] }
-iced_native = { path = "../../native" }
-iced_futures = { path = "../../futures" }
-reqwest = "0.10"
+iced.workspace = true
+iced.features = ["tokio", "sipper"]
+
+[dependencies.reqwest]
+version = "0.12"
+features = ["stream"]
