@@ -1,34 +1,54 @@
-use super::{KeyCode, ModifiersState};
+use crate::SmolStr;
+use crate::keyboard::key;
+use crate::keyboard::{Key, Location, Modifiers};
 
 /// A keyboard event.
 ///
 /// _**Note:** This type is largely incomplete! If you need to track
 /// additional events, feel free to [open an issue] and share your use case!_
 ///
-/// [open an issue]: https://github.com/hecrj/iced/issues
-#[derive(Debug, Clone, Copy, PartialEq)]
+/// [open an issue]: https://github.com/iced-rs/iced/issues
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Event {
     /// A keyboard key was pressed.
     KeyPressed {
-        /// The key identifier
-        key_code: KeyCode,
+        /// The key pressed.
+        key: Key,
 
-        /// The state of the modifier keys
-        modifiers: ModifiersState,
+        /// The key pressed with all keyboard modifiers applied, except Ctrl.
+        modified_key: Key,
+
+        /// The physical key pressed.
+        physical_key: key::Physical,
+
+        /// The location of the key.
+        location: Location,
+
+        /// The state of the modifier keys.
+        modifiers: Modifiers,
+
+        /// The text produced by the key press, if any.
+        text: Option<SmolStr>,
     },
 
     /// A keyboard key was released.
     KeyReleased {
-        /// The key identifier
-        key_code: KeyCode,
+        /// The key released.
+        key: Key,
 
-        /// The state of the modifier keys
-        modifiers: ModifiersState,
+        /// The key released with all keyboard modifiers applied, except Ctrl.
+        modified_key: Key,
+
+        /// The physical key released.
+        physical_key: key::Physical,
+
+        /// The location of the key.
+        location: Location,
+
+        /// The state of the modifier keys.
+        modifiers: Modifiers,
     },
 
-    /// A unicode character was received.
-    CharacterReceived(char),
-
     /// The keyboard modifiers have changed.
-    ModifiersChanged(ModifiersState),
+    ModifiersChanged(Modifiers),
 }
