@@ -1,8 +1,13 @@
-//! Reuse basic mouse types.
+//! Handle mouse events.
+pub mod click;
+
 mod button;
+mod cursor;
 mod event;
 mod interaction;
 
 pub use button::Button;
+pub use click::Click;
+pub use cursor::Cursor;
 pub use event::{Event, ScrollDelta};
 pub use interaction::Interaction;
