@@ -1,3 +1,5 @@
+use crate::Point;
+
 use super::Button;
 
 /// A mouse event.
@@ -5,7 +7,7 @@ use super::Button;
 /// _**Note:** This type is largely incomplete! If you need to track
 /// additional events, feel free to [open an issue] and share your use case!_
 ///
-/// [open an issue]: https://github.com/hecrj/iced/issues
+/// [open an issue]: https://github.com/iced-rs/iced/issues
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Event {
     /// The mouse cursor entered the window.
@@ -16,11 +18,8 @@ pub enum Event {
 
     /// The mouse cursor was moved
     CursorMoved {
-        /// The X coordinate of the mouse position
-        x: f32,
-
-        /// The Y coordinate of the mouse position
-        y: f32,
+        /// The new position of the mouse cursor
+        position: Point,
     },
 
     /// A mouse button was pressed.
