@@ -1,6 +1,7 @@
-use crate::{Point, Size, Vector};
+use crate::alignment;
+use crate::{Padding, Point, Radians, Size, Vector};
 
-/// A rectangle.
+/// An axis-aligned rectangle.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub struct Rectangle<T = f32> {
     /// X coordinate of the top-left corner.
@@ -16,14 +17,29 @@ pub struct Rectangle<T = f32> {
     pub height: T,
 }
 
+impl<T> Rectangle<T>
+where
+    T: Default,
+{
+    /// Creates a new [`Rectangle`] with its top-left corner at the origin
+    /// and with the provided [`Size`].
+    pub fn with_size(size: Size<T>) -> Self {
+        Self {
+            x: T::default(),
+            y: T::default(),
+            width: size.width,
+            height: size.height,
+        }
+    }
+}
+
 impl Rectangle<f32> {
+    /// A rectangle starting at [`Point::ORIGIN`] with infinite width and height.
+    pub const INFINITE: Self = Self::new(Point::ORIGIN, Size::INFINITY);
+
     /// Creates a new [`Rectangle`] with its top-left corner in the given
     /// [`Point`] and with the provided [`Size`].
-    ///
-    /// [`Rectangle`]: struct.Rectangle.html
-    /// [`Point`]: struct.Point.html
-    /// [`Size`]: struct.Size.html
-    pub fn new(top_left: Point, size: Size) -> Self {
+    pub const fn new(top_left: Point, size: Size) -> Self {
         Self {
             x: top_left.x,
             y: top_left.y,
@@ -32,75 +48,151 @@ impl Rectangle<f32> {
         }
     }
 
-    /// Creates a new [`Rectangle`] with its top-left corner at the origin
-    /// and with the provided [`Size`].
-    ///
-    /// [`Rectangle`]: struct.Rectangle.html
-    /// [`Size`]: struct.Size.html
-    pub fn with_size(size: Size) -> Self {
+    /// Creates a new square [`Rectangle`] with the center at the origin and
+    /// with the given radius.
+    pub fn with_radius(radius: f32) -> Self {
         Self {
-            x: 0.0,
-            y: 0.0,
-            width: size.width,
-            height: size.height,
+            x: -radius,
+            y: -radius,
+            width: radius * 2.0,
+            height: radius * 2.0,
         }
     }
 
+    /// Creates a new axis-aligned [`Rectangle`] from the given vertices; returning the
+    /// rotation in [`Radians`] that must be applied to the axis-aligned [`Rectangle`]
+    /// to obtain the desired result.
+    pub fn with_vertices(
+        top_left: Point,
+        top_right: Point,
+        bottom_left: Point,
+    ) -> (Rectangle, Radians) {
+        let width = (top_right.x - top_left.x).hypot(top_right.y - top_left.y);
+
+        let height =
+            (bottom_left.x - top_left.x).hypot(bottom_left.y - top_left.y);
+
+        let rotation =
+            (top_right.y - top_left.y).atan2(top_right.x - top_left.x);
+
+        let rotation = if rotation < 0.0 {
+            2.0 * std::f32::consts::PI + rotation
+        } else {
+            rotation
+        };
+
+        let position = {
+            let center = Point::new(
+                (top_right.x + bottom_left.x) / 2.0,
+                (top_right.y + bottom_left.y) / 2.0,
+            );
+
+            let rotation = -rotation - std::f32::consts::PI * 2.0;
+
+            Point::new(
+                center.x + (top_left.x - center.x) * rotation.cos()
+                    - (top_left.y - center.y) * rotation.sin(),
+                center.y
+                    + (top_left.x - center.x) * rotation.sin()
+                    + (top_left.y - center.y) * rotation.cos(),
+            )
+        };
+
+        (
+            Rectangle::new(position, Size::new(width, height)),
+            Radians(rotation),
+        )
+    }
+
     /// Returns the [`Point`] at the center of the [`Rectangle`].
-    ///
-    /// [`Point`]: struct.Point.html
-    /// [`Rectangle`]: struct.Rectangle.html
     pub fn center(&self) -> Point {
         Point::new(self.center_x(), self.center_y())
     }
 
     /// Returns the X coordinate of the [`Point`] at the center of the
     /// [`Rectangle`].
-    ///
-    /// [`Point`]: struct.Point.html
-    /// [`Rectangle`]: struct.Rectangle.html
     pub fn center_x(&self) -> f32 {
         self.x + self.width / 2.0
     }
 
     /// Returns the Y coordinate of the [`Point`] at the center of the
     /// [`Rectangle`].
-    ///
-    /// [`Point`]: struct.Point.html
-    /// [`Rectangle`]: struct.Rectangle.html
     pub fn center_y(&self) -> f32 {
         self.y + self.height / 2.0
     }
 
     /// Returns the position of the top left corner of the [`Rectangle`].
-    ///
-    /// [`Rectangle`]: struct.Rectangle.html
     pub fn position(&self) -> Point {
         Point::new(self.x, self.y)
     }
 
     /// Returns the [`Size`] of the [`Rectangle`].
-    ///
-    /// [`Size`]: struct.Size.html
-    /// [`Rectangle`]: struct.Rectangle.html
     pub fn size(&self) -> Size {
         Size::new(self.width, self.height)
     }
 
+    /// Returns the area of the [`Rectangle`].
+    pub fn area(&self) -> f32 {
+        self.width * self.height
+    }
+
     /// Returns true if the given [`Point`] is contained in the [`Rectangle`].
-    ///
-    /// [`Point`]: struct.Point.html
-    /// [`Rectangle`]: struct.Rectangle.html
+    /// Excludes the right and bottom edges.
     pub fn contains(&self, point: Point) -> bool {
         self.x <= point.x
-            && point.x <= self.x + self.width
+            && point.x < self.x + self.width
             && self.y <= point.y
-            && point.y <= self.y + self.height
+            && point.y < self.y + self.height
+    }
+
+    /// Returns the minimum distance from the given [`Point`] to any of the edges
+    /// of the [`Rectangle`].
+    pub fn distance(&self, point: Point) -> f32 {
+        let center = self.center();
+
+        let distance_x =
+            ((point.x - center.x).abs() - self.width / 2.0).max(0.0);
+
+        let distance_y =
+            ((point.y - center.y).abs() - self.height / 2.0).max(0.0);
+
+        distance_x.hypot(distance_y)
+    }
+
+    /// Computes the offset that must be applied to the [`Rectangle`] to be placed
+    /// inside the given `container`.
+    pub fn offset(&self, container: &Rectangle) -> Vector {
+        let Some(intersection) = self.intersection(container) else {
+            return Vector::ZERO;
+        };
+
+        let left = intersection.x - self.x;
+        let top = intersection.y - self.y;
+
+        Vector::new(
+            if left > 0.0 {
+                left
+            } else {
+                intersection.x + intersection.width - self.x - self.width
+            },
+            if top > 0.0 {
+                top
+            } else {
+                intersection.y + intersection.height - self.y - self.height
+            },
+        )
+    }
+
+    /// Returns true if the current [`Rectangle`] is within the given
+    /// `container`. Includes the right and bottom edges.
+    pub fn is_within(&self, container: &Rectangle) -> bool {
+        self.x >= container.x
+            && self.y >= container.y
+            && self.x + self.width <= container.x + container.width
+            && self.y + self.height <= container.y + container.height
     }
 
     /// Computes the intersection with the given [`Rectangle`].
-    ///
-    /// [`Rectangle`]: struct.Rectangle.html
     pub fn intersection(
         &self,
         other: &Rectangle<f32>,
@@ -126,16 +218,123 @@ impl Rectangle<f32> {
         }
     }
 
-    /// Snaps the [`Rectangle`] to __unsigned__ integer coordinates.
-    ///
-    /// [`Rectangle`]: struct.Rectangle.html
-    pub fn snap(self) -> Rectangle<u32> {
+    /// Returns whether the [`Rectangle`] intersects with the given one.
+    pub fn intersects(&self, other: &Self) -> bool {
+        self.intersection(other).is_some()
+    }
+
+    /// Computes the union with the given [`Rectangle`].
+    pub fn union(&self, other: &Self) -> Self {
+        let x = self.x.min(other.x);
+        let y = self.y.min(other.y);
+
+        let lower_right_x = (self.x + self.width).max(other.x + other.width);
+        let lower_right_y = (self.y + self.height).max(other.y + other.height);
+
+        let width = lower_right_x - x;
+        let height = lower_right_y - y;
+
         Rectangle {
-            x: self.x as u32,
-            y: self.y as u32,
-            width: self.width.ceil() as u32,
-            height: self.height.ceil() as u32,
+            x,
+            y,
+            width,
+            height,
         }
+    }
+
+    /// Snaps the [`Rectangle`] to __unsigned__ integer coordinates.
+    pub fn snap(self) -> Option<Rectangle<u32>> {
+        let top_left = self.position().snap();
+        let bottom_right = (self.position() + Vector::from(self.size())).snap();
+
+        let width = bottom_right.x.checked_sub(top_left.x)?;
+        let height = bottom_right.y.checked_sub(top_left.y)?;
+
+        if width < 1 || height < 1 {
+            return None;
+        }
+
+        Some(Rectangle {
+            x: top_left.x,
+            y: top_left.y,
+            width,
+            height,
+        })
+    }
+
+    /// Expands the [`Rectangle`] a given amount.
+    pub fn expand(self, padding: impl Into<Padding>) -> Self {
+        let padding = padding.into();
+
+        Self {
+            x: self.x - padding.left,
+            y: self.y - padding.top,
+            width: self.width + padding.horizontal(),
+            height: self.height + padding.vertical(),
+        }
+    }
+
+    /// Shrinks the [`Rectangle`] a given amount.
+    pub fn shrink(self, padding: impl Into<Padding>) -> Self {
+        let padding = padding.into();
+
+        Self {
+            x: self.x + padding.left,
+            y: self.y + padding.top,
+            width: self.width - padding.horizontal(),
+            height: self.height - padding.vertical(),
+        }
+    }
+
+    /// Rotates the [`Rectangle`] and returns the smallest [`Rectangle`]
+    /// containing it.
+    pub fn rotate(self, rotation: Radians) -> Self {
+        let size = self.size().rotate(rotation);
+        let position = Point::new(
+            self.center_x() - size.width / 2.0,
+            self.center_y() - size.height / 2.0,
+        );
+
+        Self::new(position, size)
+    }
+
+    /// Scales the [`Rectangle`] without changing its position, effectively
+    /// "zooming" it.
+    pub fn zoom(self, zoom: f32) -> Self {
+        Self {
+            x: self.x - (self.width * (zoom - 1.0)) / 2.0,
+            y: self.y - (self.height * (zoom - 1.0)) / 2.0,
+            width: self.width * zoom,
+            height: self.height * zoom,
+        }
+    }
+
+    /// Returns the top-left position to render an object of the given [`Size`].
+    /// inside the [`Rectangle`] that is anchored to the edge or corner
+    /// defined by the alignment arguments.
+    pub fn anchor(
+        &self,
+        size: Size,
+        align_x: impl Into<alignment::Horizontal>,
+        align_y: impl Into<alignment::Vertical>,
+    ) -> Point {
+        let x = match align_x.into() {
+            alignment::Horizontal::Left => self.x,
+            alignment::Horizontal::Center => {
+                self.x + (self.width - size.width) / 2.0
+            }
+            alignment::Horizontal::Right => self.x + self.width - size.width,
+        };
+
+        let y = match align_y.into() {
+            alignment::Vertical::Top => self.y,
+            alignment::Vertical::Center => {
+                self.y + (self.height - size.height) / 2.0
+            }
+            alignment::Vertical::Bottom => self.y + self.height - size.height,
+        };
+
+        Point::new(x, y)
     }
 }
 
@@ -144,8 +343,8 @@ impl std::ops::Mul<f32> for Rectangle<f32> {
 
     fn mul(self, scale: f32) -> Self {
         Self {
-            x: self.x as f32 * scale,
-            y: self.y as f32 * scale,
+            x: self.x * scale,
+            y: self.y * scale,
             width: self.width * scale,
             height: self.height * scale,
         }
@@ -174,6 +373,37 @@ where
             x: self.x + translation.x,
             y: self.y + translation.y,
             ..self
+        }
+    }
+}
+
+impl<T> std::ops::Sub<Vector<T>> for Rectangle<T>
+where
+    T: std::ops::Sub<Output = T>,
+{
+    type Output = Rectangle<T>;
+
+    fn sub(self, translation: Vector<T>) -> Self {
+        Rectangle {
+            x: self.x - translation.x,
+            y: self.y - translation.y,
+            ..self
+        }
+    }
+}
+
+impl<T> std::ops::Mul<Vector<T>> for Rectangle<T>
+where
+    T: std::ops::Mul<Output = T> + Copy,
+{
+    type Output = Rectangle<T>;
+
+    fn mul(self, scale: Vector<T>) -> Self {
+        Rectangle {
+            x: self.x * scale.x,
+            y: self.y * scale.y,
+            width: self.width * scale.x,
+            height: self.height * scale.y,
         }
     }
 }
