@@ -1,57 +1,67 @@
-use iced::{button, Align, Button, Column, Element, Sandbox, Settings, Text};
+use iced::Center;
+use iced::widget::{Column, button, column, text};
 
-pub fn main() {
-    Counter::run(Settings::default())
+pub fn main() -> iced::Result {
+    iced::run(Counter::update, Counter::view)
 }
 
 #[derive(Default)]
 struct Counter {
-    value: i32,
-    increment_button: button::State,
-    decrement_button: button::State,
+    value: i64,
 }
 
 #[derive(Debug, Clone, Copy)]
 enum Message {
-    IncrementPressed,
-    DecrementPressed,
+    Increment,
+    Decrement,
 }
 
-impl Sandbox for Counter {
-    type Message = Message;
-
-    fn new() -> Self {
-        Self::default()
-    }
-
-    fn title(&self) -> String {
-        String::from("Counter - Iced")
-    }
-
+impl Counter {
     fn update(&mut self, message: Message) {
         match message {
-            Message::IncrementPressed => {
+            Message::Increment => {
                 self.value += 1;
             }
-            Message::DecrementPressed => {
+            Message::Decrement => {
                 self.value -= 1;
             }
         }
     }
 
-    fn view(&mut self) -> Element<Message> {
-        Column::new()
-            .padding(20)
-            .align_items(Align::Center)
-            .push(
-                Button::new(&mut self.increment_button, Text::new("Increment"))
-                    .on_press(Message::IncrementPressed),
-            )
-            .push(Text::new(self.value.to_string()).size(50))
-            .push(
-                Button::new(&mut self.decrement_button, Text::new("Decrement"))
-                    .on_press(Message::DecrementPressed),
-            )
-            .into()
+    fn view(&self) -> Column<Message> {
+        column![
+            button("Increment").on_press(Message::Increment),
+            text(self.value).size(50),
+            button("Decrement").on_press(Message::Decrement)
+        ]
+        .padding(20)
+        .align_x(Center)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use iced_test::{Error, simulator};
+
+    #[test]
+    fn it_counts() -> Result<(), Error> {
+        let mut counter = Counter { value: 0 };
+        let mut ui = simulator(counter.view());
+
+        let _ = ui.click("Increment")?;
+        let _ = ui.click("Increment")?;
+        let _ = ui.click("Decrement")?;
+
+        for message in ui.into_messages() {
+            counter.update(message);
+        }
+
+        assert_eq!(counter.value, 1);
+
+        let mut ui = simulator(counter.view());
+        assert!(ui.find("1").is_ok(), "Counter should display 1!");
+
+        Ok(())
     }
 }
