@@ -1,87 +1,83 @@
-use iced::{
-    executor, Align, Application, Checkbox, Column, Command, Container,
-    Element, Length, Settings, Subscription, Text,
-};
+use iced::event::{self, Event};
+use iced::widget::{Column, button, center, checkbox, text};
+use iced::window;
+use iced::{Center, Element, Fill, Subscription, Task};
 
-pub fn main() {
-    Events::run(Settings::default())
+pub fn main() -> iced::Result {
+    iced::application(Events::default, Events::update, Events::view)
+        .subscription(Events::subscription)
+        .exit_on_close_request(false)
+        .run()
 }
 
 #[derive(Debug, Default)]
 struct Events {
-    last: Vec<iced_native::Event>,
+    last: Vec<Event>,
     enabled: bool,
 }
 
 #[derive(Debug, Clone)]
 enum Message {
-    EventOccurred(iced_native::Event),
+    EventOccurred(Event),
     Toggled(bool),
+    Exit,
 }
 
-impl Application for Events {
-    type Executor = executor::Default;
-    type Message = Message;
-    type Flags = ();
-
-    fn new(_flags: ()) -> (Events, Command<Message>) {
-        (Events::default(), Command::none())
-    }
-
-    fn title(&self) -> String {
-        String::from("Events - Iced")
-    }
-
-    fn update(&mut self, message: Message) -> Command<Message> {
+impl Events {
+    fn update(&mut self, message: Message) -> Task<Message> {
         match message {
-            Message::EventOccurred(event) => {
+            Message::EventOccurred(event) if self.enabled => {
                 self.last.push(event);
 
                 if self.last.len() > 5 {
                     let _ = self.last.remove(0);
                 }
+
+                Task::none()
+            }
+            Message::EventOccurred(event) => {
+                if let Event::Window(window::Event::CloseRequested) = event {
+                    window::get_latest().and_then(window::close)
+                } else {
+                    Task::none()
+                }
             }
             Message::Toggled(enabled) => {
                 self.enabled = enabled;
+
+                Task::none()
             }
-        };
-
-        Command::none()
-    }
-
-    fn subscription(&self) -> Subscription<Message> {
-        if self.enabled {
-            iced_native::subscription::events().map(Message::EventOccurred)
-        } else {
-            Subscription::none()
+            Message::Exit => window::get_latest().and_then(window::close),
         }
     }
 
-    fn view(&mut self) -> Element<Message> {
-        let events = self.last.iter().fold(
-            Column::new().spacing(10),
-            |column, event| {
-                column.push(Text::new(format!("{:?}", event)).size(40))
-            },
+    fn subscription(&self) -> Subscription<Message> {
+        event::listen().map(Message::EventOccurred)
+    }
+
+    fn view(&self) -> Element<Message> {
+        let events = Column::with_children(
+            self.last
+                .iter()
+                .map(|event| text!("{event:?}").size(40))
+                .map(Element::from),
         );
 
-        let toggle = Checkbox::new(
-            self.enabled,
-            "Listen to runtime events",
-            Message::Toggled,
-        );
+        let toggle = checkbox("Listen to runtime events", self.enabled)
+            .on_toggle(Message::Toggled);
+
+        let exit = button(text("Exit").width(Fill).align_x(Center))
+            .width(100)
+            .padding(10)
+            .on_press(Message::Exit);
 
         let content = Column::new()
-            .align_items(Align::Center)
+            .align_x(Center)
             .spacing(20)
             .push(events)
-            .push(toggle);
+            .push(toggle)
+            .push(exit);
 
-        Container::new(content)
-            .width(Length::Fill)
-            .height(Length::Fill)
-            .center_x()
-            .center_y()
-            .into()
+        center(content).into()
     }
 }
