@@ -1,7 +1,8 @@
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
 pub enum Preset {
     Custom,
-    XKCD,
+    #[default]
+    Xkcd,
     Glider,
     SmallExploder,
     Exploder,
@@ -14,7 +15,7 @@ pub enum Preset {
 
 pub static ALL: &[Preset] = &[
     Preset::Custom,
-    Preset::XKCD,
+    Preset::Xkcd,
     Preset::Glider,
     Preset::SmallExploder,
     Preset::Exploder,
@@ -30,7 +31,7 @@ impl Preset {
         #[rustfmt::skip]
         let cells = match self {
             Preset::Custom => vec![],
-            Preset::XKCD => vec![
+            Preset::Xkcd => vec![
                 "  xxx  ",
                 "  x x  ",
                 "  x x  ",
@@ -114,12 +115,6 @@ impl Preset {
     }
 }
 
-impl Default for Preset {
-    fn default() -> Preset {
-        Preset::XKCD
-    }
-}
-
 impl std::fmt::Display for Preset {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(
@@ -127,7 +122,7 @@ impl std::fmt::Display for Preset {
             "{}",
             match self {
                 Preset::Custom => "Custom",
-                Preset::XKCD => "xkcd #2293",
+                Preset::Xkcd => "xkcd #2293",
                 Preset::Glider => "Glider",
                 Preset::SmallExploder => "Small Exploder",
                 Preset::Exploder => "Exploder",
