@@ -1,24 +1,23 @@
 use iced_wgpu::Renderer;
-use iced_winit::{
-    slider, Align, Color, Column, Command, Element, Length, Program, Row,
-    Slider, Text,
-};
+use iced_widget::{bottom, column, row, slider, text, text_input};
+use iced_winit::core::{Color, Element, Theme};
 
 pub struct Controls {
     background_color: Color,
-    sliders: [slider::State; 3],
+    input: String,
 }
 
 #[derive(Debug, Clone)]
 pub enum Message {
     BackgroundColorChanged(Color),
+    InputChanged(String),
 }
 
 impl Controls {
     pub fn new() -> Controls {
         Controls {
             background_color: Color::BLACK,
-            sliders: Default::default(),
+            input: String::default(),
         }
     }
 
@@ -27,79 +26,58 @@ impl Controls {
     }
 }
 
-impl Program for Controls {
-    type Renderer = Renderer;
-    type Message = Message;
-
-    fn update(&mut self, message: Message) -> Command<Message> {
+impl Controls {
+    pub fn update(&mut self, message: Message) {
         match message {
             Message::BackgroundColorChanged(color) => {
                 self.background_color = color;
             }
+            Message::InputChanged(input) => {
+                self.input = input;
+            }
         }
-
-        Command::none()
     }
 
-    fn view(&mut self) -> Element<Message, Renderer> {
-        let [r, g, b] = &mut self.sliders;
+    pub fn view(&self) -> Element<Message, Theme, Renderer> {
         let background_color = self.background_color;
 
-        let sliders = Row::new()
-            .width(Length::Units(500))
-            .spacing(20)
-            .push(
-                Slider::new(r, 0.0..=1.0, background_color.r, move |r| {
-                    Message::BackgroundColorChanged(Color {
-                        r,
-                        ..background_color
-                    })
+        let sliders = row![
+            slider(0.0..=1.0, background_color.r, move |r| {
+                Message::BackgroundColorChanged(Color {
+                    r,
+                    ..background_color
                 })
-                .step(0.01),
-            )
-            .push(
-                Slider::new(g, 0.0..=1.0, background_color.g, move |g| {
-                    Message::BackgroundColorChanged(Color {
-                        g,
-                        ..background_color
-                    })
+            })
+            .step(0.01),
+            slider(0.0..=1.0, background_color.g, move |g| {
+                Message::BackgroundColorChanged(Color {
+                    g,
+                    ..background_color
                 })
-                .step(0.01),
-            )
-            .push(
-                Slider::new(b, 0.0..=1.0, background_color.b, move |b| {
-                    Message::BackgroundColorChanged(Color {
-                        b,
-                        ..background_color
-                    })
+            })
+            .step(0.01),
+            slider(0.0..=1.0, background_color.b, move |b| {
+                Message::BackgroundColorChanged(Color {
+                    b,
+                    ..background_color
                 })
-                .step(0.01),
-            );
+            })
+            .step(0.01),
+        ]
+        .width(500)
+        .spacing(20);
 
-        Row::new()
-            .width(Length::Fill)
-            .height(Length::Fill)
-            .align_items(Align::End)
-            .push(
-                Column::new()
-                    .width(Length::Fill)
-                    .align_items(Align::End)
-                    .push(
-                        Column::new()
-                            .padding(10)
-                            .spacing(10)
-                            .push(
-                                Text::new("Background color")
-                                    .color(Color::WHITE),
-                            )
-                            .push(sliders)
-                            .push(
-                                Text::new(format!("{:?}", background_color))
-                                    .size(14)
-                                    .color(Color::WHITE),
-                            ),
-                    ),
-            )
-            .into()
+        bottom(
+            column![
+                text("Background color").color(Color::WHITE),
+                text!("{background_color:?}").size(14).color(Color::WHITE),
+                sliders,
+                text_input("Type something...", &self.input)
+                    .on_input(Message::InputChanged),
+            ]
+            .spacing(10),
+        )
+        .padding(10)
+        .into()
     }
 }
