@@ -1,17 +1,13 @@
-use iced::{
-    pick_list, scrollable, Align, Container, Element, Length, PickList,
-    Sandbox, Scrollable, Settings, Space, Text,
-};
+use iced::widget::{column, pick_list, scrollable, vertical_space};
+use iced::{Center, Element, Fill};
 
-pub fn main() {
-    Example::run(Settings::default())
+pub fn main() -> iced::Result {
+    iced::run(Example::update, Example::view)
 }
 
 #[derive(Default)]
 struct Example {
-    scroll: scrollable::State,
-    pick_list: pick_list::State<Language>,
-    selected_language: Language,
+    selected_language: Option<Language>,
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -19,54 +15,40 @@ enum Message {
     LanguageSelected(Language),
 }
 
-impl Sandbox for Example {
-    type Message = Message;
-
-    fn new() -> Self {
-        Self::default()
-    }
-
-    fn title(&self) -> String {
-        String::from("Pick list - Iced")
-    }
-
+impl Example {
     fn update(&mut self, message: Message) {
         match message {
             Message::LanguageSelected(language) => {
-                self.selected_language = language;
+                self.selected_language = Some(language);
             }
         }
     }
 
-    fn view(&mut self) -> Element<Message> {
-        let pick_list = PickList::new(
-            &mut self.pick_list,
+    fn view(&self) -> Element<Message> {
+        let pick_list = pick_list(
             &Language::ALL[..],
-            Some(self.selected_language),
+            self.selected_language,
             Message::LanguageSelected,
-        );
+        )
+        .placeholder("Choose a language...");
 
-        let mut content = Scrollable::new(&mut self.scroll)
-            .width(Length::Fill)
-            .align_items(Align::Center)
-            .spacing(10)
-            .push(Space::with_height(Length::Units(600)))
-            .push(Text::new("Which is your favorite language?"))
-            .push(pick_list);
+        let content = column![
+            vertical_space().height(600),
+            "Which is your favorite language?",
+            pick_list,
+            vertical_space().height(600),
+        ]
+        .width(Fill)
+        .align_x(Center)
+        .spacing(10);
 
-        content = content.push(Space::with_height(Length::Units(600)));
-
-        Container::new(content)
-            .width(Length::Fill)
-            .height(Length::Fill)
-            .center_x()
-            .center_y()
-            .into()
+        scrollable(content).into()
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub enum Language {
+    #[default]
     Rust,
     Elm,
     Ruby,
@@ -86,12 +68,6 @@ impl Language {
         Language::Javascript,
         Language::Other,
     ];
-}
-
-impl Default for Language {
-    fn default() -> Language {
-        Language::Rust
-    }
 }
 
 impl std::fmt::Display for Language {
