@@ -6,19 +6,29 @@
 //! Inspired by the example found in the MDN docs[1].
 //!
 //! [1]: https://developer.mozilla.org/en-US/docs/Web/API/Canvas_API/Tutorial/Basic_animations#An_animated_solar_system
+use iced::mouse;
+use iced::widget::canvas::stroke::{self, Stroke};
+use iced::widget::canvas::{Geometry, Path};
+use iced::widget::{canvas, image};
+use iced::window;
 use iced::{
-    canvas::{self, Cursor, Path, Stroke},
-    executor, time, window, Application, Canvas, Color, Command, Element,
-    Length, Point, Rectangle, Settings, Size, Subscription, Vector,
+    Color, Element, Fill, Point, Rectangle, Renderer, Size, Subscription,
+    Theme, Vector,
 };
 
 use std::time::Instant;
 
-pub fn main() {
-    SolarSystem::run(Settings {
-        antialiasing: true,
-        ..Settings::default()
-    })
+pub fn main() -> iced::Result {
+    tracing_subscriber::fmt::init();
+
+    iced::application::timed(
+        SolarSystem::new,
+        SolarSystem::update,
+        SolarSystem::subscription,
+        SolarSystem::view,
+    )
+    .theme(SolarSystem::theme)
+    .run()
 }
 
 struct SolarSystem {
@@ -27,55 +37,44 @@ struct SolarSystem {
 
 #[derive(Debug, Clone, Copy)]
 enum Message {
-    Tick(Instant),
+    Tick,
 }
 
-impl Application for SolarSystem {
-    type Executor = executor::Default;
-    type Message = Message;
-    type Flags = ();
-
-    fn new(_flags: ()) -> (Self, Command<Message>) {
-        (
-            SolarSystem {
-                state: State::new(),
-            },
-            Command::none(),
-        )
+impl SolarSystem {
+    fn new() -> Self {
+        Self {
+            state: State::new(),
+        }
     }
 
-    fn title(&self) -> String {
-        String::from("Solar system - Iced")
-    }
-
-    fn update(&mut self, message: Message) -> Command<Message> {
+    fn update(&mut self, message: Message, now: Instant) {
         match message {
-            Message::Tick(instant) => {
-                self.state.update(instant);
+            Message::Tick => {
+                self.state.update(now);
             }
         }
+    }
 
-        Command::none()
+    fn view(&self) -> Element<Message> {
+        canvas(&self.state).width(Fill).height(Fill).into()
+    }
+
+    fn theme(&self) -> Theme {
+        Theme::Moonfly
     }
 
     fn subscription(&self) -> Subscription<Message> {
-        time::every(std::time::Duration::from_millis(10))
-            .map(|instant| Message::Tick(instant))
-    }
-
-    fn view(&mut self) -> Element<Message> {
-        Canvas::new(&mut self.state)
-            .width(Length::Fill)
-            .height(Length::Fill)
-            .into()
+        window::frames().map(|_| Message::Tick)
     }
 }
 
 #[derive(Debug)]
 struct State {
+    sun: image::Handle,
+    earth: image::Handle,
+    moon: image::Handle,
     space_cache: canvas::Cache,
     system_cache: canvas::Cache,
-    cursor_position: Point,
     start: Instant,
     now: Instant,
     stars: Vec<(Point, f32)>,
@@ -90,24 +89,33 @@ impl State {
 
     pub fn new() -> State {
         let now = Instant::now();
-        let (width, height) = window::Settings::default().size;
+        let size = window::Settings::default().size;
 
         State {
-            space_cache: Default::default(),
-            system_cache: Default::default(),
-            cursor_position: Point::ORIGIN,
+            sun: image::Handle::from_bytes(
+                include_bytes!("../assets/sun.png").as_slice(),
+            ),
+            earth: image::Handle::from_bytes(
+                include_bytes!("../assets/earth.png").as_slice(),
+            ),
+            moon: image::Handle::from_bytes(
+                include_bytes!("../assets/moon.png").as_slice(),
+            ),
+            space_cache: canvas::Cache::default(),
+            system_cache: canvas::Cache::default(),
             start: now,
             now,
-            stars: Self::generate_stars(width, height),
+            stars: Self::generate_stars(size.width, size.height),
         }
     }
 
     pub fn update(&mut self, now: Instant) {
+        self.start = self.start.min(now);
         self.now = now;
         self.system_cache.clear();
     }
 
-    fn generate_stars(width: u32, height: u32) -> Vec<(Point, f32)> {
+    fn generate_stars(width: f32, height: f32) -> Vec<(Point, f32)> {
         use rand::Rng;
 
         let mut rng = rand::thread_rng();
@@ -116,16 +124,10 @@ impl State {
             .map(|_| {
                 (
                     Point::new(
-                        rng.gen_range(
-                            -(width as f32) / 2.0,
-                            width as f32 / 2.0,
-                        ),
-                        rng.gen_range(
-                            -(height as f32) / 2.0,
-                            height as f32 / 2.0,
-                        ),
+                        rng.gen_range((-width / 2.0)..(width / 2.0)),
+                        rng.gen_range((-height / 2.0)..(height / 2.0)),
                     ),
-                    rng.gen_range(0.5, 1.0),
+                    rng.gen_range(0.5..1.0),
                 )
             })
             .collect()
@@ -133,40 +135,51 @@ impl State {
 }
 
 impl<Message> canvas::Program<Message> for State {
+    type State = ();
+
     fn draw(
         &self,
+        _state: &Self::State,
+        renderer: &Renderer,
+        _theme: &Theme,
         bounds: Rectangle,
-        _cursor: Cursor,
-    ) -> Vec<canvas::Geometry> {
+        _cursor: mouse::Cursor,
+    ) -> Vec<Geometry> {
         use std::f32::consts::PI;
 
-        let background = self.space_cache.draw(bounds.size(), |frame| {
-            let space = Path::rectangle(Point::new(0.0, 0.0), frame.size());
+        let background =
+            self.space_cache.draw(renderer, bounds.size(), |frame| {
+                frame.fill_rectangle(Point::ORIGIN, frame.size(), Color::BLACK);
 
-            let stars = Path::new(|path| {
-                for (p, size) in &self.stars {
-                    path.rectangle(*p, Size::new(*size, *size));
-                }
+                let stars = Path::new(|path| {
+                    for (p, size) in &self.stars {
+                        path.rectangle(*p, Size::new(*size, *size));
+                    }
+                });
+
+                frame.translate(frame.center() - Point::ORIGIN);
+                frame.fill(&stars, Color::WHITE);
             });
 
-            frame.fill(&space, Color::BLACK);
-
-            frame.translate(frame.center() - Point::ORIGIN);
-            frame.fill(&stars, Color::WHITE);
-        });
-
-        let system = self.system_cache.draw(bounds.size(), |frame| {
+        let system = self.system_cache.draw(renderer, bounds.size(), |frame| {
             let center = frame.center();
+            frame.translate(Vector::new(center.x, center.y));
 
-            let sun = Path::circle(center, Self::SUN_RADIUS);
-            let orbit = Path::circle(center, Self::ORBIT_RADIUS);
+            frame.draw_image(
+                Rectangle::with_radius(Self::SUN_RADIUS),
+                &self.sun,
+            );
 
-            frame.fill(&sun, Color::from_rgb8(0xF9, 0xD7, 0x1C));
+            let orbit = Path::circle(Point::ORIGIN, Self::ORBIT_RADIUS);
             frame.stroke(
                 &orbit,
                 Stroke {
+                    style: stroke::Style::Solid(Color::WHITE.scale_alpha(0.1)),
                     width: 1.0,
-                    color: Color::from_rgba8(0, 153, 255, 0.1),
+                    line_dash: canvas::LineDash {
+                        offset: 0,
+                        segments: &[3.0, 6.0],
+                    },
                     ..Stroke::default()
                 },
             );
@@ -175,38 +188,21 @@ impl<Message> canvas::Program<Message> for State {
             let rotation = (2.0 * PI / 60.0) * elapsed.as_secs() as f32
                 + (2.0 * PI / 60_000.0) * elapsed.subsec_millis() as f32;
 
-            frame.with_save(|frame| {
-                frame.translate(Vector::new(center.x, center.y));
-                frame.rotate(rotation);
-                frame.translate(Vector::new(Self::ORBIT_RADIUS, 0.0));
+            frame.rotate(rotation);
+            frame.translate(Vector::new(Self::ORBIT_RADIUS, 0.0));
 
-                let earth = Path::circle(Point::ORIGIN, Self::EARTH_RADIUS);
-                let shadow = Path::rectangle(
-                    Point::new(0.0, -Self::EARTH_RADIUS),
-                    Size::new(
-                        Self::EARTH_RADIUS * 4.0,
-                        Self::EARTH_RADIUS * 2.0,
-                    ),
-                );
+            frame.draw_image(
+                Rectangle::with_radius(Self::EARTH_RADIUS),
+                canvas::Image::new(&self.earth).rotation(-rotation * 20.0),
+            );
 
-                frame.fill(&earth, Color::from_rgb8(0x6B, 0x93, 0xD6));
+            frame.rotate(rotation * 10.0);
+            frame.translate(Vector::new(0.0, Self::MOON_DISTANCE));
 
-                frame.with_save(|frame| {
-                    frame.rotate(rotation * 10.0);
-                    frame.translate(Vector::new(0.0, Self::MOON_DISTANCE));
-
-                    let moon = Path::circle(Point::ORIGIN, Self::MOON_RADIUS);
-                    frame.fill(&moon, Color::WHITE);
-                });
-
-                frame.fill(
-                    &shadow,
-                    Color {
-                        a: 0.7,
-                        ..Color::BLACK
-                    },
-                );
-            });
+            frame.draw_image(
+                Rectangle::with_radius(Self::MOON_RADIUS),
+                &self.moon,
+            );
         });
 
         vec![background, system]
