@@ -1,37 +1,35 @@
-use crate::{BoxFuture, Subscription};
+use crate::subscription::{Event, Hasher, Recipe};
+use crate::{BoxFuture, MaybeSend};
 
-use futures::{channel::mpsc, sink::Sink};
-use std::{collections::HashMap, marker::PhantomData};
+use futures::channel::mpsc;
+use futures::sink::{Sink, SinkExt};
+use rustc_hash::FxHashMap;
+
+use std::hash::Hasher as _;
 
 /// A registry of subscription streams.
 ///
 /// If you have an application that continuously returns a [`Subscription`],
 /// you can use a [`Tracker`] to keep track of the different recipes and keep
 /// its executions alive.
-#[derive(Debug)]
-pub struct Tracker<Hasher, Event> {
-    subscriptions: HashMap<u64, Execution<Event>>,
-    _hasher: PhantomData<Hasher>,
+///
+/// [`Subscription`]: crate::Subscription
+#[derive(Debug, Default)]
+pub struct Tracker {
+    subscriptions: FxHashMap<u64, Execution>,
 }
 
 #[derive(Debug)]
-pub struct Execution<Event> {
+pub struct Execution {
     _cancel: futures::channel::oneshot::Sender<()>,
     listener: Option<futures::channel::mpsc::Sender<Event>>,
 }
 
-impl<Hasher, Event> Tracker<Hasher, Event>
-where
-    Hasher: std::hash::Hasher + Default,
-    Event: 'static + Send + Clone,
-{
+impl Tracker {
     /// Creates a new empty [`Tracker`].
-    ///
-    /// [`Tracker`]: struct.Tracker.html
     pub fn new() -> Self {
         Self {
-            subscriptions: HashMap::new(),
-            _hasher: PhantomData,
+            subscriptions: FxHashMap::default(),
         }
     }
 
@@ -44,35 +42,32 @@ where
     /// method:
     ///
     /// - If the provided [`Subscription`] contains a new [`Recipe`] that is
-    /// currently not being run, it will spawn a new stream and keep it alive.
+    ///   currently not being run, it will spawn a new stream and keep it alive.
     /// - On the other hand, if a [`Recipe`] is currently in execution and the
-    /// provided [`Subscription`] does not contain it anymore, then the
-    /// [`Tracker`] will close and drop the relevant stream.
+    ///   provided [`Subscription`] does not contain it anymore, then the
+    ///   [`Tracker`] will close and drop the relevant stream.
     ///
     /// It returns a list of futures that need to be spawned to materialize
     /// the [`Tracker`] changes.
     ///
-    /// [`Tracker`]: struct.Tracker.html
-    /// [`Subscription`]: struct.Subscription.html
-    /// [`Recipe`]: trait.Recipe.html
+    /// [`Recipe`]: crate::subscription::Recipe
+    /// [`Subscription`]: crate::Subscription
     pub fn update<Message, Receiver>(
         &mut self,
-        subscription: Subscription<Hasher, Event, Message>,
+        recipes: impl Iterator<Item = Box<dyn Recipe<Output = Message>>>,
         receiver: Receiver,
     ) -> Vec<BoxFuture<()>>
     where
-        Message: 'static + Send,
+        Message: 'static + MaybeSend,
         Receiver: 'static
             + Sink<Message, Error = mpsc::SendError>
             + Unpin
-            + Send
+            + MaybeSend
             + Clone,
     {
-        use futures::{future::FutureExt, stream::StreamExt};
+        use futures::stream::StreamExt;
 
         let mut futures: Vec<BoxFuture<()>> = Vec::new();
-
-        let recipes = subscription.recipes();
         let mut alive = std::collections::HashSet::new();
 
         for recipe in recipes {
@@ -89,19 +84,29 @@ where
                 continue;
             }
 
-            let (cancel, cancelled) = futures::channel::oneshot::channel();
+            let (cancel, mut canceled) = futures::channel::oneshot::channel();
 
             // TODO: Use bus if/when it supports async
             let (event_sender, event_receiver) =
                 futures::channel::mpsc::channel(100);
 
-            let stream = recipe.stream(event_receiver.boxed());
+            let mut receiver = receiver.clone();
+            let mut stream = recipe.stream(event_receiver.boxed());
 
-            let future = futures::future::select(
-                cancelled,
-                stream.map(Ok).forward(receiver.clone()),
-            )
-            .map(|_| ());
+            let future = async move {
+                loop {
+                    let select =
+                        futures::future::select(&mut canceled, stream.next());
+
+                    match select.await {
+                        futures::future::Either::Left(_)
+                        | futures::future::Either::Right((None, _)) => break,
+                        futures::future::Either::Right((Some(message), _)) => {
+                            let _ = receiver.send(message).await;
+                        }
+                    }
+                }
+            };
 
             let _ = self.subscriptions.insert(
                 id,
@@ -118,7 +123,7 @@ where
             futures.push(Box::pin(future));
         }
 
-        self.subscriptions.retain(|id, _| alive.contains(&id));
+        self.subscriptions.retain(|id, _| alive.contains(id));
 
         futures
     }
@@ -132,16 +137,15 @@ where
     /// This method publishes the given event to all the subscription streams
     /// currently open.
     ///
-    /// [`Recipe::stream`]: trait.Recipe.html#tymethod.stream
+    /// [`Recipe::stream`]: crate::subscription::Recipe::stream
     pub fn broadcast(&mut self, event: Event) {
         self.subscriptions
             .values_mut()
             .filter_map(|connection| connection.listener.as_mut())
             .for_each(|listener| {
                 if let Err(error) = listener.try_send(event.clone()) {
-                    log::error!(
-                        "Error sending event to subscription: {:?}",
-                        error
+                    log::warn!(
+                        "Error sending event to subscription: {error:?}"
                     );
                 }
             });
