@@ -1,359 +1,144 @@
-//! Organize rendering primitives into a flattened list of layers.
-use crate::image;
-use crate::svg;
-use crate::triangle;
-use crate::{
-    Background, Font, HorizontalAlignment, Point, Primitive, Rectangle, Size,
-    Vector, VerticalAlignment, Viewport,
-};
+//! Draw and stack layers of graphical primitives.
+use crate::core::{Rectangle, Transformation};
 
-/// A group of primitives that should be clipped together.
-#[derive(Debug, Clone)]
-pub struct Layer<'a> {
-    /// The clipping bounds of the [`Layer`].
-    ///
-    /// [`Layer`]: struct.Layer.html
-    pub bounds: Rectangle,
-
-    /// The quads of the [`Layer`].
-    ///
-    /// [`Layer`]: struct.Layer.html
-    pub quads: Vec<Quad>,
-
-    /// The triangle meshes of the [`Layer`].
-    ///
-    /// [`Layer`]: struct.Layer.html
-    pub meshes: Vec<Mesh<'a>>,
-
-    /// The text of the [`Layer`].
-    ///
-    /// [`Layer`]: struct.Layer.html
-    pub text: Vec<Text<'a>>,
-
-    /// The images of the [`Layer`].
-    ///
-    /// [`Layer`]: struct.Layer.html
-    pub images: Vec<Image>,
-}
-
-impl<'a> Layer<'a> {
-    /// Creates a new [`Layer`] with the given clipping bounds.
-    ///
-    /// [`Layer`]: struct.Layer.html
-    pub fn new(bounds: Rectangle) -> Self {
-        Self {
-            bounds,
-            quads: Vec::new(),
-            meshes: Vec::new(),
-            text: Vec::new(),
-            images: Vec::new(),
-        }
-    }
-
-    /// Creates a new [`Layer`] for the provided overlay text.
-    ///
-    /// This can be useful for displaying debug information.
-    ///
-    /// [`Layer`]: struct.Layer.html
-    pub fn overlay(lines: &'a [impl AsRef<str>], viewport: &Viewport) -> Self {
-        let mut overlay =
-            Layer::new(Rectangle::with_size(viewport.logical_size()));
-
-        for (i, line) in lines.iter().enumerate() {
-            let text = Text {
-                content: line.as_ref(),
-                bounds: Rectangle::new(
-                    Point::new(11.0, 11.0 + 25.0 * i as f32),
-                    Size::INFINITY,
-                ),
-                color: [0.9, 0.9, 0.9, 1.0],
-                size: 20.0,
-                font: Font::Default,
-                horizontal_alignment: HorizontalAlignment::Left,
-                vertical_alignment: VerticalAlignment::Top,
-            };
-
-            overlay.text.push(text);
-
-            overlay.text.push(Text {
-                bounds: text.bounds + Vector::new(-1.0, -1.0),
-                color: [0.0, 0.0, 0.0, 1.0],
-                ..text
-            });
-        }
-
-        overlay
-    }
-
-    /// Distributes the given [`Primitive`] and generates a list of layers based
-    /// on its contents.
-    ///
-    /// [`Primitive`]: ../enum.Primitive.html
-    pub fn generate(
-        primitive: &'a Primitive,
-        viewport: &Viewport,
-    ) -> Vec<Self> {
-        let first_layer =
-            Layer::new(Rectangle::with_size(viewport.logical_size()));
-
-        let mut layers = vec![first_layer];
-
-        Self::process_primitive(&mut layers, Vector::new(0.0, 0.0), primitive);
-
-        layers
-    }
-
-    fn process_primitive(
-        layers: &mut Vec<Self>,
-        translation: Vector,
-        primitive: &'a Primitive,
-    ) {
-        match primitive {
-            Primitive::None => {}
-            Primitive::Group { primitives } => {
-                // TODO: Inspect a bit and regroup (?)
-                for primitive in primitives {
-                    Self::process_primitive(layers, translation, primitive)
-                }
-            }
-            Primitive::Text {
-                content,
-                bounds,
-                size,
-                color,
-                font,
-                horizontal_alignment,
-                vertical_alignment,
-            } => {
-                let layer = layers.last_mut().unwrap();
-
-                layer.text.push(Text {
-                    content,
-                    bounds: *bounds + translation,
-                    size: *size,
-                    color: color.into_linear(),
-                    font: *font,
-                    horizontal_alignment: *horizontal_alignment,
-                    vertical_alignment: *vertical_alignment,
-                });
-            }
-            Primitive::Quad {
-                bounds,
-                background,
-                border_radius,
-                border_width,
-                border_color,
-            } => {
-                let layer = layers.last_mut().unwrap();
-
-                // TODO: Move some of these computations to the GPU (?)
-                layer.quads.push(Quad {
-                    position: [
-                        bounds.x + translation.x,
-                        bounds.y + translation.y,
-                    ],
-                    size: [bounds.width, bounds.height],
-                    color: match background {
-                        Background::Color(color) => color.into_linear(),
-                    },
-                    border_radius: *border_radius as f32,
-                    border_width: *border_width as f32,
-                    border_color: border_color.into_linear(),
-                });
-            }
-            Primitive::Mesh2D { buffers, size } => {
-                let layer = layers.last_mut().unwrap();
-
-                let bounds = Rectangle::new(
-                    Point::new(translation.x, translation.y),
-                    *size,
-                );
-
-                // Only draw visible content
-                if let Some(clip_bounds) = layer.bounds.intersection(&bounds) {
-                    layer.meshes.push(Mesh {
-                        origin: Point::new(translation.x, translation.y),
-                        buffers,
-                        clip_bounds,
-                    });
-                }
-            }
-            Primitive::Clip {
-                bounds,
-                offset,
-                content,
-            } => {
-                let layer = layers.last_mut().unwrap();
-                let translated_bounds = *bounds + translation;
-
-                // Only draw visible content
-                if let Some(clip_bounds) =
-                    layer.bounds.intersection(&translated_bounds)
-                {
-                    let clip_layer = Layer::new(clip_bounds);
-                    let new_layer = Layer::new(layer.bounds);
-
-                    layers.push(clip_layer);
-                    Self::process_primitive(
-                        layers,
-                        translation
-                            - Vector::new(offset.x as f32, offset.y as f32),
-                        content,
-                    );
-                    layers.push(new_layer);
-                }
-            }
-            Primitive::Translate {
-                translation: new_translation,
-                content,
-            } => {
-                Self::process_primitive(
-                    layers,
-                    translation + *new_translation,
-                    &content,
-                );
-            }
-            Primitive::Cached { cache } => {
-                Self::process_primitive(layers, translation, &cache);
-            }
-            Primitive::Image { handle, bounds } => {
-                let layer = layers.last_mut().unwrap();
-
-                layer.images.push(Image::Raster {
-                    handle: handle.clone(),
-                    bounds: *bounds + translation,
-                });
-            }
-            Primitive::Svg { handle, bounds } => {
-                let layer = layers.last_mut().unwrap();
-
-                layer.images.push(Image::Vector {
-                    handle: handle.clone(),
-                    bounds: *bounds + translation,
-                });
-            }
-        }
-    }
-}
-
-/// A colored rectangle with a border.
+/// A layer of graphical primitives.
 ///
-/// This type can be directly uploaded to GPU memory.
-#[derive(Debug, Clone, Copy)]
-#[repr(C)]
-pub struct Quad {
-    /// The position of the [`Quad`].
-    ///
-    /// [`Quad`]: struct.Quad.html
-    pub position: [f32; 2],
+/// Layers normally dictate a set of primitives that are
+/// rendered in a specific order.
+pub trait Layer: Default {
+    /// Creates a new [`Layer`] with the given bounds.
+    fn with_bounds(bounds: Rectangle) -> Self;
 
-    /// The size of the [`Quad`].
+    /// Flushes and settles any pending group of primitives in the [`Layer`].
     ///
-    /// [`Quad`]: struct.Quad.html
-    pub size: [f32; 2],
+    /// This will be called when a [`Layer`] is finished. It allows layers to efficiently
+    /// record primitives together and defer grouping until the end.
+    fn flush(&mut self);
 
-    /// The color of the [`Quad`], in __linear RGB__.
-    ///
-    /// [`Quad`]: struct.Quad.html
-    pub color: [f32; 4],
+    /// Resizes the [`Layer`] to the given bounds.
+    fn resize(&mut self, bounds: Rectangle);
 
-    /// The border color of the [`Quad`], in __linear RGB__.
-    ///
-    /// [`Quad`]: struct.Quad.html
-    pub border_color: [f32; 4],
-
-    /// The border radius of the [`Quad`].
-    ///
-    /// [`Quad`]: struct.Quad.html
-    pub border_radius: f32,
-
-    /// The border width of the [`Quad`].
-    ///
-    /// [`Quad`]: struct.Quad.html
-    pub border_width: f32,
+    /// Clears all the layers contents and resets its bounds.
+    fn reset(&mut self);
 }
 
-/// A mesh of triangles.
-#[derive(Debug, Clone, Copy)]
-pub struct Mesh<'a> {
-    /// The origin of the vertices of the [`Mesh`].
-    ///
-    /// [`Mesh`]: struct.Mesh.html
-    pub origin: Point,
-
-    /// The vertex and index buffers of the [`Mesh`].
-    ///
-    /// [`Mesh`]: struct.Mesh.html
-    pub buffers: &'a triangle::Mesh2D,
-
-    /// The clipping bounds of the [`Mesh`].
-    ///
-    /// [`Mesh`]: struct.Mesh.html
-    pub clip_bounds: Rectangle<f32>,
+/// A stack of layers used for drawing.
+#[derive(Debug)]
+pub struct Stack<T: Layer> {
+    layers: Vec<T>,
+    transformations: Vec<Transformation>,
+    previous: Vec<usize>,
+    current: usize,
+    active_count: usize,
 }
 
-/// A paragraph of text.
-#[derive(Debug, Clone, Copy)]
-pub struct Text<'a> {
-    /// The content of the [`Text`].
-    ///
-    /// [`Text`]: struct.Text.html
-    pub content: &'a str,
+impl<T: Layer> Stack<T> {
+    /// Creates a new empty [`Stack`].
+    pub fn new() -> Self {
+        Self {
+            layers: vec![T::default()],
+            transformations: vec![Transformation::IDENTITY],
+            previous: vec![],
+            current: 0,
+            active_count: 1,
+        }
+    }
 
-    /// The layout bounds of the [`Text`].
-    ///
-    /// [`Text`]: struct.Text.html
-    pub bounds: Rectangle,
+    /// Returns a mutable reference to the current [`Layer`] of the [`Stack`], together with
+    /// the current [`Transformation`].
+    #[inline]
+    pub fn current_mut(&mut self) -> (&mut T, Transformation) {
+        let transformation = self.transformation();
 
-    /// The color of the [`Text`], in __linear RGB_.
-    ///
-    /// [`Text`]: struct.Text.html
-    pub color: [f32; 4],
+        (&mut self.layers[self.current], transformation)
+    }
 
-    /// The size of the [`Text`].
-    ///
-    /// [`Text`]: struct.Text.html
-    pub size: f32,
+    /// Returns the current [`Transformation`] of the [`Stack`].
+    #[inline]
+    pub fn transformation(&self) -> Transformation {
+        self.transformations.last().copied().unwrap()
+    }
 
-    /// The font of the [`Text`].
-    ///
-    /// [`Text`]: struct.Text.html
-    pub font: Font,
+    /// Pushes a new clipping region in the [`Stack`]; creating a new layer in the
+    /// process.
+    pub fn push_clip(&mut self, bounds: Rectangle) {
+        self.previous.push(self.current);
 
-    /// The horizontal alignment of the [`Text`].
-    ///
-    /// [`Text`]: struct.Text.html
-    pub horizontal_alignment: HorizontalAlignment,
+        self.current = self.active_count;
+        self.active_count += 1;
 
-    /// The vertical alignment of the [`Text`].
+        let bounds = bounds * self.transformation();
+
+        if self.current == self.layers.len() {
+            self.layers.push(T::with_bounds(bounds));
+        } else {
+            self.layers[self.current].resize(bounds);
+        }
+    }
+
+    /// Pops the current clipping region from the [`Stack`] and restores the previous one.
     ///
-    /// [`Text`]: struct.Text.html
-    pub vertical_alignment: VerticalAlignment,
+    /// The current layer will be recorded for drawing.
+    pub fn pop_clip(&mut self) {
+        self.flush();
+
+        self.current = self.previous.pop().unwrap();
+    }
+
+    /// Pushes a new [`Transformation`] in the [`Stack`].
+    ///
+    /// Future drawing operations will be affected by this new [`Transformation`] until
+    /// it is popped using [`pop_transformation`].
+    ///
+    /// [`pop_transformation`]: Self::pop_transformation
+    pub fn push_transformation(&mut self, transformation: Transformation) {
+        self.transformations
+            .push(self.transformation() * transformation);
+    }
+
+    /// Pops the current [`Transformation`] in the [`Stack`].
+    pub fn pop_transformation(&mut self) {
+        let _ = self.transformations.pop();
+    }
+
+    /// Returns an iterator over mutable references to the layers in the [`Stack`].
+    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut T> {
+        self.flush();
+
+        self.layers[..self.active_count].iter_mut()
+    }
+
+    /// Returns an iterator over immutable references to the layers in the [`Stack`].
+    pub fn iter(&self) -> impl Iterator<Item = &T> {
+        self.layers[..self.active_count].iter()
+    }
+
+    /// Returns the slice of layers in the [`Stack`].
+    pub fn as_slice(&self) -> &[T] {
+        &self.layers[..self.active_count]
+    }
+
+    /// Flushes and settles any primitives in the current layer of the [`Stack`].
+    pub fn flush(&mut self) {
+        self.layers[self.current].flush();
+    }
+
+    /// Clears the layers of the [`Stack`], allowing reuse.
+    ///
+    /// This will normally keep layer allocations for future drawing operations.
+    pub fn clear(&mut self) {
+        for layer in self.layers[..self.active_count].iter_mut() {
+            layer.reset();
+        }
+
+        self.current = 0;
+        self.active_count = 1;
+        self.previous.clear();
+    }
 }
 
-/// A raster or vector image.
-#[derive(Debug, Clone)]
-pub enum Image {
-    /// A raster image.
-    Raster {
-        /// The handle of a raster image.
-        handle: image::Handle,
-
-        /// The bounds of the image.
-        bounds: Rectangle,
-    },
-    /// A vector image.
-    Vector {
-        /// The handle of a vector image.
-        handle: svg::Handle,
-
-        /// The bounds of the image.
-        bounds: Rectangle,
-    },
+impl<T: Layer> Default for Stack<T> {
+    fn default() -> Self {
+        Self::new()
+    }
 }
-
-#[allow(unsafe_code)]
-unsafe impl bytemuck::Zeroable for Quad {}
-
-#[allow(unsafe_code)]
-unsafe impl bytemuck::Pod for Quad {}
