@@ -1,327 +1,523 @@
+//! Create and run iced applications step by step.
+//!
+//! # Example
+//! ```no_run,standalone_crate
+//! use iced::widget::{button, column, text, Column};
+//! use iced::Theme;
+//!
+//! pub fn main() -> iced::Result {
+//!     iced::application(u64::default, update, view)
+//!         .theme(|_| Theme::Dark)
+//!         .centered()
+//!         .run()
+//! }
+//!
+//! #[derive(Debug, Clone)]
+//! enum Message {
+//!     Increment,
+//! }
+//!
+//! fn update(value: &mut u64, message: Message) {
+//!     match message {
+//!         Message::Increment => *value += 1,
+//!     }
+//! }
+//!
+//! fn view(value: &u64) -> Column<Message> {
+//!     column![
+//!         text(value),
+//!         button("+").on_press(Message::Increment),
+//!     ]
+//! }
+//! ```
+use crate::program::{self, Program};
+use crate::shell;
+use crate::theme;
+use crate::window;
 use crate::{
-    window, Color, Command, Element, Executor, Settings, Subscription,
+    Element, Executor, Font, Result, Settings, Size, Subscription, Task,
 };
 
-/// An interactive cross-platform application.
+use std::borrow::Cow;
+
+pub mod timed;
+
+pub use timed::timed;
+
+/// Creates an iced [`Application`] given its boot, update, and view logic.
 ///
-/// This trait is the main entrypoint of Iced. Once implemented, you can run
-/// your GUI application by simply calling [`run`](#method.run).
+/// # Example
+/// ```no_run,standalone_crate
+/// use iced::widget::{button, column, text, Column};
 ///
-/// - On native platforms, it will run in its own window.
-/// - On the web, it will take control of the `<title>` and the `<body>` of the
-///   document.
-///
-/// An [`Application`] can execute asynchronous actions by returning a
-/// [`Command`](struct.Command.html) in some of its methods. If
-/// you do not intend to perform any background work in your program, the
-/// [`Sandbox`](trait.Sandbox.html) trait offers a simplified interface.
-///
-/// When using an [`Application`] with the `debug` feature enabled, a debug view
-/// can be toggled by pressing `F12`.
-///
-/// [`Application`]: trait.Application.html
-///
-/// # Examples
-/// [The repository has a bunch of examples] that use the [`Application`] trait:
-///
-/// - [`clock`], an application that uses the [`Canvas`] widget to draw a clock
-/// and its hands to display the current time.
-/// - [`download_progress`], a basic application that asynchronously downloads
-/// a dummy file of 100 MB and tracks the download progress.
-/// - [`events`], a log of native events displayed using a conditional
-/// [`Subscription`].
-/// - [`pokedex`], an application that displays a random Pokédex entry (sprite
-/// included!) by using the [PokéAPI].
-/// - [`solar_system`], an animated solar system drawn using the [`Canvas`] widget
-/// and showcasing how to compose different transforms.
-/// - [`stopwatch`], a watch with start/stop and reset buttons showcasing how
-/// to listen to time.
-/// - [`todos`], a todos tracker inspired by [TodoMVC].
-///
-/// [The repository has a bunch of examples]: https://github.com/hecrj/iced/tree/0.1/examples
-/// [`clock`]: https://github.com/hecrj/iced/tree/0.1/examples/clock
-/// [`download_progress`]: https://github.com/hecrj/iced/tree/0.1/examples/download_progress
-/// [`events`]: https://github.com/hecrj/iced/tree/0.1/examples/events
-/// [`pokedex`]: https://github.com/hecrj/iced/tree/0.1/examples/pokedex
-/// [`solar_system`]: https://github.com/hecrj/iced/tree/0.1/examples/solar_system
-/// [`stopwatch`]: https://github.com/hecrj/iced/tree/0.1/examples/stopwatch
-/// [`todos`]: https://github.com/hecrj/iced/tree/0.1/examples/todos
-/// [`Canvas`]: widget/canvas/struct.Canvas.html
-/// [PokéAPI]: https://pokeapi.co/
-/// [`Subscription`]: type.Subscription.html
-/// [TodoMVC]: http://todomvc.com/
-///
-/// ## A simple "Hello, world!"
-///
-/// If you just want to get started, here is a simple [`Application`] that
-/// says "Hello, world!":
-///
-/// ```no_run
-/// use iced::{executor, Application, Command, Element, Settings, Text};
-///
-/// pub fn main() {
-///     Hello::run(Settings::default())
+/// pub fn main() -> iced::Result {
+///     iced::application(u64::default, update, view).run()
 /// }
 ///
-/// struct Hello;
+/// #[derive(Debug, Clone)]
+/// enum Message {
+///     Increment,
+/// }
 ///
-/// impl Application for Hello {
-///     type Executor = executor::Null;
-///     type Message = ();
-///     type Flags = ();
-///
-///     fn new(_flags: ()) -> (Hello, Command<Self::Message>) {
-///         (Hello, Command::none())
+/// fn update(value: &mut u64, message: Message) {
+///     match message {
+///         Message::Increment => *value += 1,
 ///     }
+/// }
 ///
-///     fn title(&self) -> String {
-///         String::from("A cool application")
-///     }
-///
-///     fn update(&mut self, _message: Self::Message) -> Command<Self::Message> {
-///         Command::none()
-///     }
-///
-///     fn view(&mut self) -> Element<Self::Message> {
-///         Text::new("Hello, world!").into()
-///     }
+/// fn view(value: &u64) -> Column<Message> {
+///     column![
+///         text(value),
+///         button("+").on_press(Message::Increment),
+///     ]
 /// }
 /// ```
-pub trait Application: Sized {
-    /// The [`Executor`] that will run commands and subscriptions.
-    ///
-    /// The [default executor] can be a good starting point!
-    ///
-    /// [`Executor`]: trait.Executor.html
-    /// [default executor]: executor/struct.Default.html
-    type Executor: Executor;
+pub fn application<State, Message, Theme, Renderer>(
+    boot: impl Boot<State, Message>,
+    update: impl Update<State, Message>,
+    view: impl for<'a> View<'a, State, Message, Theme, Renderer>,
+) -> Application<impl Program<State = State, Message = Message, Theme = Theme>>
+where
+    State: 'static,
+    Message: program::Message + 'static,
+    Theme: Default + theme::Base,
+    Renderer: program::Renderer,
+{
+    use std::marker::PhantomData;
 
-    /// The type of __messages__ your [`Application`] will produce.
-    ///
-    /// [`Application`]: trait.Application.html
-    type Message: std::fmt::Debug + Send;
-
-    /// The data needed to initialize your [`Application`].
-    ///
-    /// [`Application`]: trait.Application.html
-    type Flags;
-
-    /// Initializes the [`Application`] with the flags provided to
-    /// [`run`] as part of the [`Settings`].
-    ///
-    /// Here is where you should return the initial state of your app.
-    ///
-    /// Additionally, you can return a [`Command`](struct.Command.html) if you
-    /// need to perform some async action in the background on startup. This is
-    /// useful if you want to load state from a file, perform an initial HTTP
-    /// request, etc.
-    ///
-    /// [`Application`]: trait.Application.html
-    /// [`run`]: #method.run.html
-    /// [`Settings`]: struct.Settings.html
-    fn new(flags: Self::Flags) -> (Self, Command<Self::Message>);
-
-    /// Returns the current title of the [`Application`].
-    ///
-    /// This title can be dynamic! The runtime will automatically update the
-    /// title of your application when necessary.
-    ///
-    /// [`Application`]: trait.Application.html
-    fn title(&self) -> String;
-
-    /// Handles a __message__ and updates the state of the [`Application`].
-    ///
-    /// This is where you define your __update logic__. All the __messages__,
-    /// produced by either user interactions or commands, will be handled by
-    /// this method.
-    ///
-    /// Any [`Command`] returned will be executed immediately in the background.
-    ///
-    /// [`Application`]: trait.Application.html
-    /// [`Command`]: struct.Command.html
-    fn update(&mut self, message: Self::Message) -> Command<Self::Message>;
-
-    /// Returns the event [`Subscription`] for the current state of the
-    /// application.
-    ///
-    /// A [`Subscription`] will be kept alive as long as you keep returning it,
-    /// and the __messages__ produced will be handled by
-    /// [`update`](#tymethod.update).
-    ///
-    /// By default, this method returns an empty [`Subscription`].
-    ///
-    /// [`Subscription`]: struct.Subscription.html
-    fn subscription(&self) -> Subscription<Self::Message> {
-        Subscription::none()
+    struct Instance<State, Message, Theme, Renderer, Boot, Update, View> {
+        boot: Boot,
+        update: Update,
+        view: View,
+        _state: PhantomData<State>,
+        _message: PhantomData<Message>,
+        _theme: PhantomData<Theme>,
+        _renderer: PhantomData<Renderer>,
     }
 
-    /// Returns the widgets to display in the [`Application`].
-    ///
-    /// These widgets can produce __messages__ based on user interaction.
-    ///
-    /// [`Application`]: trait.Application.html
-    fn view(&mut self) -> Element<'_, Self::Message>;
+    impl<State, Message, Theme, Renderer, Boot, Update, View> Program
+        for Instance<State, Message, Theme, Renderer, Boot, Update, View>
+    where
+        Message: program::Message + 'static,
+        Theme: Default + theme::Base,
+        Renderer: program::Renderer,
+        Boot: self::Boot<State, Message>,
+        Update: self::Update<State, Message>,
+        View: for<'a> self::View<'a, State, Message, Theme, Renderer>,
+    {
+        type State = State;
+        type Message = Message;
+        type Theme = Theme;
+        type Renderer = Renderer;
+        type Executor = iced_futures::backend::default::Executor;
 
-    /// Returns the current [`Application`] mode.
-    ///
-    /// The runtime will automatically transition your application if a new mode
-    /// is returned.
-    ///
-    /// Currently, the mode only has an effect in native platforms.
-    ///
-    /// By default, an application will run in windowed mode.
-    ///
-    /// [`Application`]: trait.Application.html
-    fn mode(&self) -> window::Mode {
-        window::Mode::Windowed
+        fn name() -> &'static str {
+            let name = std::any::type_name::<State>();
+
+            name.split("::").next().unwrap_or("a_cool_application")
+        }
+
+        fn boot(&self) -> (State, Task<Message>) {
+            self.boot.boot()
+        }
+
+        fn update(
+            &self,
+            state: &mut Self::State,
+            message: Self::Message,
+        ) -> Task<Self::Message> {
+            self.update.update(state, message)
+        }
+
+        fn view<'a>(
+            &self,
+            state: &'a Self::State,
+            _window: window::Id,
+        ) -> Element<'a, Self::Message, Self::Theme, Self::Renderer> {
+            self.view.view(state)
+        }
     }
 
-    /// Returns the background color of the [`Application`].
-    ///
-    /// By default, it returns [`Color::WHITE`].
-    ///
-    /// [`Application`]: trait.Application.html
-    /// [`Color::WHITE`]: struct.Color.html#const.WHITE
-    fn background_color(&self) -> Color {
-        Color::WHITE
+    Application {
+        raw: Instance {
+            boot,
+            update,
+            view,
+            _state: PhantomData,
+            _message: PhantomData,
+            _theme: PhantomData,
+            _renderer: PhantomData,
+        },
+        settings: Settings::default(),
+        window: window::Settings::default(),
     }
+}
 
-    /// Returns the scale factor of the [`Application`].
-    ///
-    /// It can be used to dynamically control the size of the UI at runtime
-    /// (i.e. zooming).
-    ///
-    /// For instance, a scale factor of `2.0` will make widgets twice as big,
-    /// while a scale factor of `0.5` will shrink them to half their size.
-    ///
-    /// By default, it returns `1.0`.
-    ///
-    /// [`Application`]: trait.Application.html
-    fn scale_factor(&self) -> f64 {
-        1.0
-    }
+/// The underlying definition and configuration of an iced application.
+///
+/// You can use this API to create and run iced applications
+/// step by step—without coupling your logic to a trait
+/// or a specific type.
+///
+/// You can create an [`Application`] with the [`application`] helper.
+#[derive(Debug)]
+pub struct Application<P: Program> {
+    raw: P,
+    settings: Settings,
+    window: window::Settings,
+}
 
+impl<P: Program> Application<P> {
     /// Runs the [`Application`].
-    ///
-    /// On native platforms, this method will take control of the current thread
-    /// and __will NOT return__.
-    ///
-    /// It should probably be that last thing you call in your `main` function.
-    ///
-    /// [`Application`]: trait.Application.html
-    fn run(settings: Settings<Self::Flags>)
+    pub fn run(self) -> Result
     where
         Self: 'static,
     {
-        #[cfg(not(target_arch = "wasm32"))]
-        {
-            let renderer_settings = crate::renderer::Settings {
-                default_font: settings.default_font,
-                default_text_size: settings.default_text_size,
-                antialiasing: if settings.antialiasing {
-                    Some(crate::renderer::settings::Antialiasing::MSAAx4)
-                } else {
-                    None
-                },
-                ..crate::renderer::Settings::default()
-            };
+        #[cfg(all(feature = "debug", not(target_arch = "wasm32")))]
+        let program = {
+            iced_debug::init(iced_debug::Metadata {
+                name: P::name(),
+                theme: None,
+                can_time_travel: cfg!(feature = "time-travel"),
+            });
 
-            crate::runtime::application::run::<
-                Instance<Self>,
-                Self::Executor,
-                crate::renderer::window::Compositor,
-            >(settings.into(), renderer_settings);
-        }
+            iced_devtools::attach(self.raw)
+        };
 
-        #[cfg(target_arch = "wasm32")]
-        <Instance<Self> as iced_web::Application>::run(settings.flags);
-    }
-}
+        #[cfg(any(not(feature = "debug"), target_arch = "wasm32"))]
+        let program = self.raw;
 
-struct Instance<A: Application>(A);
-
-#[cfg(not(target_arch = "wasm32"))]
-impl<A> iced_winit::Program for Instance<A>
-where
-    A: Application,
-{
-    type Renderer = crate::renderer::Renderer;
-    type Message = A::Message;
-
-    fn update(&mut self, message: Self::Message) -> Command<Self::Message> {
-        self.0.update(message)
+        Ok(shell::run(program, self.settings, Some(self.window))?)
     }
 
-    fn view(&mut self) -> Element<'_, Self::Message> {
-        self.0.view()
-    }
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-impl<A> crate::runtime::Application for Instance<A>
-where
-    A: Application,
-{
-    type Flags = A::Flags;
-
-    fn new(flags: Self::Flags) -> (Self, Command<A::Message>) {
-        let (app, command) = A::new(flags);
-
-        (Instance(app), command)
+    /// Sets the [`Settings`] that will be used to run the [`Application`].
+    pub fn settings(self, settings: Settings) -> Self {
+        Self { settings, ..self }
     }
 
-    fn title(&self) -> String {
-        self.0.title()
-    }
-
-    fn mode(&self) -> iced_winit::Mode {
-        match self.0.mode() {
-            window::Mode::Windowed => iced_winit::Mode::Windowed,
-            window::Mode::Fullscreen => iced_winit::Mode::Fullscreen,
+    /// Sets the [`Settings::antialiasing`] of the [`Application`].
+    pub fn antialiasing(self, antialiasing: bool) -> Self {
+        Self {
+            settings: Settings {
+                antialiasing,
+                ..self.settings
+            },
+            ..self
         }
     }
 
-    fn subscription(&self) -> Subscription<Self::Message> {
-        self.0.subscription()
+    /// Sets the default [`Font`] of the [`Application`].
+    pub fn default_font(self, default_font: Font) -> Self {
+        Self {
+            settings: Settings {
+                default_font,
+                ..self.settings
+            },
+            ..self
+        }
     }
 
-    fn background_color(&self) -> Color {
-        self.0.background_color()
+    /// Adds a font to the list of fonts that will be loaded at the start of the [`Application`].
+    pub fn font(mut self, font: impl Into<Cow<'static, [u8]>>) -> Self {
+        self.settings.fonts.push(font.into());
+        self
     }
 
-    fn scale_factor(&self) -> f64 {
-        self.0.scale_factor()
+    /// Sets the [`window::Settings`] of the [`Application`].
+    ///
+    /// Overwrites any previous [`window::Settings`].
+    pub fn window(self, window: window::Settings) -> Self {
+        Self { window, ..self }
+    }
+
+    /// Sets the [`window::Settings::position`] to [`window::Position::Centered`] in the [`Application`].
+    pub fn centered(self) -> Self {
+        Self {
+            window: window::Settings {
+                position: window::Position::Centered,
+                ..self.window
+            },
+            ..self
+        }
+    }
+
+    /// Sets the [`window::Settings::exit_on_close_request`] of the [`Application`].
+    pub fn exit_on_close_request(self, exit_on_close_request: bool) -> Self {
+        Self {
+            window: window::Settings {
+                exit_on_close_request,
+                ..self.window
+            },
+            ..self
+        }
+    }
+
+    /// Sets the [`window::Settings::size`] of the [`Application`].
+    pub fn window_size(self, size: impl Into<Size>) -> Self {
+        Self {
+            window: window::Settings {
+                size: size.into(),
+                ..self.window
+            },
+            ..self
+        }
+    }
+
+    /// Sets the [`window::Settings::transparent`] of the [`Application`].
+    pub fn transparent(self, transparent: bool) -> Self {
+        Self {
+            window: window::Settings {
+                transparent,
+                ..self.window
+            },
+            ..self
+        }
+    }
+
+    /// Sets the [`window::Settings::resizable`] of the [`Application`].
+    pub fn resizable(self, resizable: bool) -> Self {
+        Self {
+            window: window::Settings {
+                resizable,
+                ..self.window
+            },
+            ..self
+        }
+    }
+
+    /// Sets the [`window::Settings::decorations`] of the [`Application`].
+    pub fn decorations(self, decorations: bool) -> Self {
+        Self {
+            window: window::Settings {
+                decorations,
+                ..self.window
+            },
+            ..self
+        }
+    }
+
+    /// Sets the [`window::Settings::position`] of the [`Application`].
+    pub fn position(self, position: window::Position) -> Self {
+        Self {
+            window: window::Settings {
+                position,
+                ..self.window
+            },
+            ..self
+        }
+    }
+
+    /// Sets the [`window::Settings::level`] of the [`Application`].
+    pub fn level(self, level: window::Level) -> Self {
+        Self {
+            window: window::Settings {
+                level,
+                ..self.window
+            },
+            ..self
+        }
+    }
+
+    /// Sets the [`Title`] of the [`Application`].
+    pub fn title(
+        self,
+        title: impl Title<P::State>,
+    ) -> Application<
+        impl Program<State = P::State, Message = P::Message, Theme = P::Theme>,
+    > {
+        Application {
+            raw: program::with_title(self.raw, move |state, _window| {
+                title.title(state)
+            }),
+            settings: self.settings,
+            window: self.window,
+        }
+    }
+
+    /// Sets the subscription logic of the [`Application`].
+    pub fn subscription(
+        self,
+        f: impl Fn(&P::State) -> Subscription<P::Message>,
+    ) -> Application<
+        impl Program<State = P::State, Message = P::Message, Theme = P::Theme>,
+    > {
+        Application {
+            raw: program::with_subscription(self.raw, f),
+            settings: self.settings,
+            window: self.window,
+        }
+    }
+
+    /// Sets the theme logic of the [`Application`].
+    pub fn theme(
+        self,
+        f: impl Fn(&P::State) -> P::Theme,
+    ) -> Application<
+        impl Program<State = P::State, Message = P::Message, Theme = P::Theme>,
+    > {
+        Application {
+            raw: program::with_theme(self.raw, move |state, _window| f(state)),
+            settings: self.settings,
+            window: self.window,
+        }
+    }
+
+    /// Sets the style logic of the [`Application`].
+    pub fn style(
+        self,
+        f: impl Fn(&P::State, &P::Theme) -> theme::Style,
+    ) -> Application<
+        impl Program<State = P::State, Message = P::Message, Theme = P::Theme>,
+    > {
+        Application {
+            raw: program::with_style(self.raw, f),
+            settings: self.settings,
+            window: self.window,
+        }
+    }
+
+    /// Sets the scale factor of the [`Application`].
+    pub fn scale_factor(
+        self,
+        f: impl Fn(&P::State) -> f64,
+    ) -> Application<
+        impl Program<State = P::State, Message = P::Message, Theme = P::Theme>,
+    > {
+        Application {
+            raw: program::with_scale_factor(self.raw, move |state, _window| {
+                f(state)
+            }),
+            settings: self.settings,
+            window: self.window,
+        }
+    }
+
+    /// Sets the executor of the [`Application`].
+    pub fn executor<E>(
+        self,
+    ) -> Application<
+        impl Program<State = P::State, Message = P::Message, Theme = P::Theme>,
+    >
+    where
+        E: Executor,
+    {
+        Application {
+            raw: program::with_executor::<P, E>(self.raw),
+            settings: self.settings,
+            window: self.window,
+        }
     }
 }
 
-#[cfg(target_arch = "wasm32")]
-impl<A> iced_web::Application for Instance<A>
+/// The logic to initialize the `State` of some [`Application`].
+///
+/// This trait is implemented for both `Fn() -> State` and
+/// `Fn() -> (State, Task<Message>)`.
+///
+/// In practice, this means that [`application`] can both take
+/// simple functions like `State::default` and more advanced ones
+/// that return a [`Task`].
+pub trait Boot<State, Message> {
+    /// Initializes the [`Application`] state.
+    fn boot(&self) -> (State, Task<Message>);
+}
+
+impl<T, C, State, Message> Boot<State, Message> for T
 where
-    A: Application,
+    T: Fn() -> C,
+    C: IntoBoot<State, Message>,
 {
-    type Executor = A::Executor;
-    type Message = A::Message;
-    type Flags = A::Flags;
-
-    fn new(flags: Self::Flags) -> (Self, Command<A::Message>) {
-        let (app, command) = A::new(flags);
-
-        (Instance(app), command)
+    fn boot(&self) -> (State, Task<Message>) {
+        self().into_boot()
     }
+}
 
-    fn title(&self) -> String {
-        self.0.title()
+/// The initial state of some [`Application`].
+pub trait IntoBoot<State, Message> {
+    /// Turns some type into the initial state of some [`Application`].
+    fn into_boot(self) -> (State, Task<Message>);
+}
+
+impl<State, Message> IntoBoot<State, Message> for State {
+    fn into_boot(self) -> (State, Task<Message>) {
+        (self, Task::none())
     }
+}
 
-    fn update(&mut self, message: Self::Message) -> Command<Self::Message> {
-        self.0.update(message)
+impl<State, Message> IntoBoot<State, Message> for (State, Task<Message>) {
+    fn into_boot(self) -> (State, Task<Message>) {
+        self
     }
+}
 
-    fn subscription(&self) -> Subscription<Self::Message> {
-        self.0.subscription()
+/// The title logic of some [`Application`].
+///
+/// This trait is implemented both for `&static str` and
+/// any closure `Fn(&State) -> String`.
+///
+/// This trait allows the [`application`] builder to take any of them.
+pub trait Title<State> {
+    /// Produces the title of the [`Application`].
+    fn title(&self, state: &State) -> String;
+}
+
+impl<State> Title<State> for &'static str {
+    fn title(&self, _state: &State) -> String {
+        self.to_string()
     }
+}
 
-    fn view(&mut self) -> Element<'_, Self::Message> {
-        self.0.view()
+impl<T, State> Title<State> for T
+where
+    T: Fn(&State) -> String,
+{
+    fn title(&self, state: &State) -> String {
+        self(state)
+    }
+}
+
+/// The update logic of some [`Application`].
+///
+/// This trait allows the [`application`] builder to take any closure that
+/// returns any `Into<Task<Message>>`.
+pub trait Update<State, Message> {
+    /// Processes the message and updates the state of the [`Application`].
+    fn update(&self, state: &mut State, message: Message) -> Task<Message>;
+}
+
+impl<State, Message> Update<State, Message> for () {
+    fn update(&self, _state: &mut State, _message: Message) -> Task<Message> {
+        Task::none()
+    }
+}
+
+impl<T, State, Message, C> Update<State, Message> for T
+where
+    T: Fn(&mut State, Message) -> C,
+    C: Into<Task<Message>>,
+{
+    fn update(&self, state: &mut State, message: Message) -> Task<Message> {
+        self(state, message).into()
+    }
+}
+
+/// The view logic of some [`Application`].
+///
+/// This trait allows the [`application`] builder to take any closure that
+/// returns any `Into<Element<'_, Message>>`.
+pub trait View<'a, State, Message, Theme, Renderer> {
+    /// Produces the widget of the [`Application`].
+    fn view(&self, state: &'a State) -> Element<'a, Message, Theme, Renderer>;
+}
+
+impl<'a, T, State, Message, Theme, Renderer, Widget>
+    View<'a, State, Message, Theme, Renderer> for T
+where
+    T: Fn(&'a State) -> Widget,
+    State: 'static,
+    Widget: Into<Element<'a, Message, Theme, Renderer>>,
+{
+    fn view(&self, state: &'a State) -> Element<'a, Message, Theme, Renderer> {
+        self(state).into()
     }
 }
