@@ -1,9 +1,8 @@
 //! Configure the window of your application in native platforms.
-mod mode;
-mod settings;
 
 pub mod icon;
 
 pub use icon::Icon;
-pub use mode::Mode;
-pub use settings::Settings;
+
+pub use crate::core::window::*;
+pub use crate::runtime::window::*;
