@@ -1,37 +1,79 @@
 //! Configure a renderer.
-pub use crate::Antialiasing;
+use crate::core::{Font, Pixels};
+use crate::graphics::{self, Antialiasing};
 
 /// The settings of a [`Renderer`].
 ///
-/// [`Renderer`]: ../struct.Renderer.html
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+/// [`Renderer`]: crate::Renderer
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub struct Settings {
-    /// The output format of the [`Renderer`].
+    /// The present mode of the [`Renderer`].
     ///
-    /// [`Renderer`]: ../struct.Renderer.html
-    pub format: wgpu::TextureFormat,
+    /// [`Renderer`]: crate::Renderer
+    pub present_mode: wgpu::PresentMode,
 
-    /// The bytes of the font that will be used by default.
-    ///
-    /// If `None` is provided, a default system font will be chosen.
-    pub default_font: Option<&'static [u8]>,
+    /// The graphics backends to use.
+    pub backends: wgpu::Backends,
+
+    /// The default [`Font`] to use.
+    pub default_font: Font,
 
     /// The default size of text.
     ///
-    /// By default, it will be set to 20.
-    pub default_text_size: u16,
+    /// By default, it will be set to `16.0`.
+    pub default_text_size: Pixels,
 
     /// The antialiasing strategy that will be used for triangle primitives.
+    ///
+    /// By default, it is `None`.
     pub antialiasing: Option<Antialiasing>,
 }
 
 impl Default for Settings {
     fn default() -> Settings {
         Settings {
-            format: wgpu::TextureFormat::Bgra8UnormSrgb,
-            default_font: None,
-            default_text_size: 20,
+            present_mode: wgpu::PresentMode::AutoVsync,
+            backends: wgpu::Backends::all(),
+            default_font: Font::default(),
+            default_text_size: Pixels(16.0),
             antialiasing: None,
         }
+    }
+}
+
+impl From<graphics::Settings> for Settings {
+    fn from(settings: graphics::Settings) -> Self {
+        Self {
+            default_font: settings.default_font,
+            default_text_size: settings.default_text_size,
+            antialiasing: settings.antialiasing,
+            ..Settings::default()
+        }
+    }
+}
+
+/// Obtains a [`wgpu::PresentMode`] from the current environment
+/// configuration, if set.
+///
+/// The value returned by this function can be changed by setting
+/// the `ICED_PRESENT_MODE` env variable. The possible values are:
+///
+/// - `vsync` → [`wgpu::PresentMode::AutoVsync`]
+/// - `no_vsync` → [`wgpu::PresentMode::AutoNoVsync`]
+/// - `immediate` → [`wgpu::PresentMode::Immediate`]
+/// - `fifo` → [`wgpu::PresentMode::Fifo`]
+/// - `fifo_relaxed` → [`wgpu::PresentMode::FifoRelaxed`]
+/// - `mailbox` → [`wgpu::PresentMode::Mailbox`]
+pub fn present_mode_from_env() -> Option<wgpu::PresentMode> {
+    let present_mode = std::env::var("ICED_PRESENT_MODE").ok()?;
+
+    match present_mode.to_lowercase().as_str() {
+        "vsync" => Some(wgpu::PresentMode::AutoVsync),
+        "no_vsync" => Some(wgpu::PresentMode::AutoNoVsync),
+        "immediate" => Some(wgpu::PresentMode::Immediate),
+        "fifo" => Some(wgpu::PresentMode::Fifo),
+        "fifo_relaxed" => Some(wgpu::PresentMode::FifoRelaxed),
+        "mailbox" => Some(wgpu::PresentMode::Mailbox),
+        _ => None,
     }
 }
