@@ -1,4 +1,5 @@
 //! Display rendering results on windows.
-mod compositor;
+pub mod compositor;
 
 pub use compositor::Compositor;
+pub use wgpu::Surface;
