@@ -1,42 +1,1496 @@
 //! A windowing shell for Iced, on top of [`winit`].
 //!
-//! ![`iced_winit` crate graph](https://github.com/hecrj/iced/blob/cae26cb7bc627f4a5b3bcf1cd023a0c552e8c65e/docs/graphs/winit.png?raw=true)
+//! ![The native path of the Iced ecosystem](https://github.com/iced-rs/iced/blob/0525d76ff94e828b7b21634fa94a747022001c83/docs/graphs/native.png?raw=true)
 //!
-//! `iced_winit` offers some convenient abstractions on top of [`iced_native`]
+//! `iced_winit` offers some convenient abstractions on top of [`iced_runtime`]
 //! to quickstart development when using [`winit`].
 //!
-//! It exposes a renderer-agnostic [`Application`] trait that can be implemented
+//! It exposes a renderer-agnostic [`Program`] trait that can be implemented
 //! and then run with a simple call. The use of this trait is optional.
 //!
 //! Additionally, a [`conversion`] module is available for users that decide to
 //! implement a custom event loop.
 //!
-//! [`iced_native`]: https://github.com/hecrj/iced/tree/master/native
+//! [`iced_runtime`]: https://github.com/iced-rs/iced/tree/0.13/runtime
 //! [`winit`]: https://github.com/rust-windowing/winit
-//! [`Application`]: trait.Application.html
-//! [`conversion`]: conversion
-#![deny(missing_docs)]
-#![deny(missing_debug_implementations)]
-#![deny(unused_results)]
-#![forbid(unsafe_code)]
-#![forbid(rust_2018_idioms)]
-
-#[doc(no_inline)]
-pub use iced_native::*;
+//! [`conversion`]: crate::conversion
+#![doc(
+    html_logo_url = "https://raw.githubusercontent.com/iced-rs/iced/9ab6923e943f784985e9ef9ca28b10278297225d/docs/logo.svg"
+)]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+pub use iced_program as program;
+pub use program::core;
+pub use program::graphics;
+pub use program::runtime;
+pub use runtime::debug;
+pub use runtime::futures;
 pub use winit;
 
-pub mod application;
+pub mod clipboard;
 pub mod conversion;
-pub mod settings;
 
-mod clipboard;
-mod mode;
+#[cfg(feature = "system")]
+pub mod system;
+
+mod error;
 mod proxy;
+mod window;
 
-pub use application::Application;
 pub use clipboard::Clipboard;
-pub use mode::Mode;
+pub use error::Error;
 pub use proxy::Proxy;
-pub use settings::Settings;
 
-pub use iced_graphics::Viewport;
+use crate::core::mouse;
+use crate::core::renderer;
+use crate::core::theme;
+use crate::core::time::Instant;
+use crate::core::widget::operation;
+use crate::core::{Point, Settings, Size};
+use crate::futures::futures::channel::mpsc;
+use crate::futures::futures::channel::oneshot;
+use crate::futures::futures::task;
+use crate::futures::futures::{Future, StreamExt};
+use crate::futures::subscription;
+use crate::futures::{Executor, Runtime};
+use crate::graphics::{Compositor, compositor};
+use crate::runtime::user_interface::{self, UserInterface};
+use crate::runtime::{Action, Task};
+
+use program::Program;
+use window::WindowManager;
+
+use rustc_hash::FxHashMap;
+use std::borrow::Cow;
+use std::mem::ManuallyDrop;
+use std::sync::Arc;
+
+/// Runs a [`Program`] with the provided settings.
+pub fn run<P>(
+    program: P,
+    settings: Settings,
+    window_settings: Option<window::Settings>,
+) -> Result<(), Error>
+where
+    P: Program + 'static,
+    P::Theme: theme::Base,
+{
+    use winit::event_loop::EventLoop;
+
+    let boot_span = debug::boot();
+
+    let graphics_settings = settings.clone().into();
+    let event_loop = EventLoop::with_user_event()
+        .build()
+        .expect("Create event loop");
+
+    let (proxy, worker) = Proxy::new(event_loop.create_proxy());
+
+    let mut runtime = {
+        let executor =
+            P::Executor::new().map_err(Error::ExecutorCreationFailed)?;
+        executor.spawn(worker);
+
+        Runtime::new(executor, proxy.clone())
+    };
+
+    let (program, task) = runtime.enter(|| program::Instance::new(program));
+    let is_daemon = window_settings.is_none();
+
+    let task = if let Some(window_settings) = window_settings {
+        let mut task = Some(task);
+
+        let (_id, open) = runtime::window::open(window_settings);
+
+        open.then(move |_| task.take().unwrap_or(Task::none()))
+    } else {
+        task
+    };
+
+    if let Some(stream) = runtime::task::into_stream(task) {
+        runtime.run(stream);
+    }
+
+    runtime.track(subscription::into_recipes(
+        runtime.enter(|| program.subscription().map(Action::Output)),
+    ));
+
+    let (event_sender, event_receiver) = mpsc::unbounded();
+    let (control_sender, control_receiver) = mpsc::unbounded();
+
+    let instance = Box::pin(run_instance::<P>(
+        program,
+        runtime,
+        proxy.clone(),
+        event_receiver,
+        control_sender,
+        is_daemon,
+        graphics_settings,
+        settings.fonts,
+    ));
+
+    let context = task::Context::from_waker(task::noop_waker_ref());
+
+    struct Runner<Message: 'static, F> {
+        instance: std::pin::Pin<Box<F>>,
+        context: task::Context<'static>,
+        id: Option<String>,
+        sender: mpsc::UnboundedSender<Event<Action<Message>>>,
+        receiver: mpsc::UnboundedReceiver<Control>,
+        error: Option<Error>,
+
+        #[cfg(target_arch = "wasm32")]
+        canvas: Option<web_sys::HtmlCanvasElement>,
+    }
+
+    let runner = Runner {
+        instance,
+        context,
+        id: settings.id,
+        sender: event_sender,
+        receiver: control_receiver,
+        error: None,
+
+        #[cfg(target_arch = "wasm32")]
+        canvas: None,
+    };
+
+    boot_span.finish();
+
+    impl<Message, F> winit::application::ApplicationHandler<Action<Message>>
+        for Runner<Message, F>
+    where
+        Message: std::fmt::Debug,
+        F: Future<Output = ()>,
+    {
+        fn resumed(
+            &mut self,
+            _event_loop: &winit::event_loop::ActiveEventLoop,
+        ) {
+        }
+
+        fn new_events(
+            &mut self,
+            event_loop: &winit::event_loop::ActiveEventLoop,
+            cause: winit::event::StartCause,
+        ) {
+            self.process_event(
+                event_loop,
+                Event::EventLoopAwakened(winit::event::Event::NewEvents(cause)),
+            );
+        }
+
+        fn window_event(
+            &mut self,
+            event_loop: &winit::event_loop::ActiveEventLoop,
+            window_id: winit::window::WindowId,
+            event: winit::event::WindowEvent,
+        ) {
+            #[cfg(target_os = "windows")]
+            let is_move_or_resize = matches!(
+                event,
+                winit::event::WindowEvent::Resized(_)
+                    | winit::event::WindowEvent::Moved(_)
+            );
+
+            self.process_event(
+                event_loop,
+                Event::EventLoopAwakened(winit::event::Event::WindowEvent {
+                    window_id,
+                    event,
+                }),
+            );
+
+            // TODO: Remove when unnecessary
+            // On Windows, we emulate an `AboutToWait` event after every `Resized` event
+            // since the event loop does not resume during resize interaction.
+            // More details: https://github.com/rust-windowing/winit/issues/3272
+            #[cfg(target_os = "windows")]
+            {
+                if is_move_or_resize {
+                    self.process_event(
+                        event_loop,
+                        Event::EventLoopAwakened(
+                            winit::event::Event::AboutToWait,
+                        ),
+                    );
+                }
+            }
+        }
+
+        fn user_event(
+            &mut self,
+            event_loop: &winit::event_loop::ActiveEventLoop,
+            action: Action<Message>,
+        ) {
+            self.process_event(
+                event_loop,
+                Event::EventLoopAwakened(winit::event::Event::UserEvent(
+                    action,
+                )),
+            );
+        }
+
+        fn received_url(
+            &mut self,
+            event_loop: &winit::event_loop::ActiveEventLoop,
+            url: String,
+        ) {
+            self.process_event(
+                event_loop,
+                Event::EventLoopAwakened(
+                    winit::event::Event::PlatformSpecific(
+                        winit::event::PlatformSpecific::MacOS(
+                            winit::event::MacOS::ReceivedUrl(url),
+                        ),
+                    ),
+                ),
+            );
+        }
+
+        fn about_to_wait(
+            &mut self,
+            event_loop: &winit::event_loop::ActiveEventLoop,
+        ) {
+            self.process_event(
+                event_loop,
+                Event::EventLoopAwakened(winit::event::Event::AboutToWait),
+            );
+        }
+    }
+
+    impl<Message, F> Runner<Message, F>
+    where
+        F: Future<Output = ()>,
+    {
+        fn process_event(
+            &mut self,
+            event_loop: &winit::event_loop::ActiveEventLoop,
+            event: Event<Action<Message>>,
+        ) {
+            if event_loop.exiting() {
+                return;
+            }
+
+            self.sender.start_send(event).expect("Send event");
+
+            loop {
+                let poll = self.instance.as_mut().poll(&mut self.context);
+
+                match poll {
+                    task::Poll::Pending => match self.receiver.try_next() {
+                        Ok(Some(control)) => match control {
+                            Control::ChangeFlow(flow) => {
+                                use winit::event_loop::ControlFlow;
+
+                                match (event_loop.control_flow(), flow) {
+                                    (
+                                        ControlFlow::WaitUntil(current),
+                                        ControlFlow::WaitUntil(new),
+                                    ) if current < new => {}
+                                    (
+                                        ControlFlow::WaitUntil(target),
+                                        ControlFlow::Wait,
+                                    ) if target > Instant::now() => {}
+                                    _ => {
+                                        event_loop.set_control_flow(flow);
+                                    }
+                                }
+                            }
+                            Control::CreateWindow {
+                                id,
+                                settings,
+                                title,
+                                monitor,
+                                on_open,
+                            } => {
+                                let exit_on_close_request =
+                                    settings.exit_on_close_request;
+
+                                let visible = settings.visible;
+
+                                #[cfg(target_arch = "wasm32")]
+                                let target =
+                                    settings.platform_specific.target.clone();
+
+                                let window_attributes =
+                                    conversion::window_attributes(
+                                        settings,
+                                        &title,
+                                        monitor
+                                            .or(event_loop.primary_monitor()),
+                                        self.id.clone(),
+                                    )
+                                    .with_visible(false);
+
+                                #[cfg(target_arch = "wasm32")]
+                                let window_attributes = {
+                                    use winit::platform::web::WindowAttributesExtWebSys;
+                                    window_attributes
+                                        .with_canvas(self.canvas.take())
+                                };
+
+                                log::info!(
+                                    "Window attributes for id `{id:#?}`: {window_attributes:#?}"
+                                );
+
+                                // On macOS, the `position` in `WindowAttributes` represents the "inner"
+                                // position of the window; while on other platforms it's the "outer" position.
+                                // We fix the inconsistency on macOS by positioning the window after creation.
+                                #[cfg(target_os = "macos")]
+                                let mut window_attributes = window_attributes;
+
+                                #[cfg(target_os = "macos")]
+                                let position =
+                                    window_attributes.position.take();
+
+                                let window = event_loop
+                                    .create_window(window_attributes)
+                                    .expect("Create window");
+
+                                #[cfg(target_os = "macos")]
+                                if let Some(position) = position {
+                                    window.set_outer_position(position);
+                                }
+
+                                #[cfg(target_arch = "wasm32")]
+                                {
+                                    use winit::platform::web::WindowExtWebSys;
+
+                                    let canvas = window
+                                        .canvas()
+                                        .expect("Get window canvas");
+
+                                    let _ = canvas.set_attribute(
+                                        "style",
+                                        "display: block; width: 100%; height: 100%",
+                                    );
+
+                                    let window = web_sys::window().unwrap();
+                                    let document = window.document().unwrap();
+                                    let body = document.body().unwrap();
+
+                                    let target = target.and_then(|target| {
+                                        body.query_selector(&format!(
+                                            "#{target}"
+                                        ))
+                                        .ok()
+                                        .unwrap_or(None)
+                                    });
+
+                                    match target {
+                                        Some(node) => {
+                                            let _ = node
+                                                .replace_with_with_node_1(
+                                                    &canvas,
+                                                )
+                                                .expect(&format!(
+                                                    "Could not replace #{}",
+                                                    node.id()
+                                                ));
+                                        }
+                                        None => {
+                                            let _ = body
+                                                .append_child(&canvas)
+                                                .expect(
+                                                "Append canvas to HTML body",
+                                            );
+                                        }
+                                    };
+                                }
+
+                                self.process_event(
+                                    event_loop,
+                                    Event::WindowCreated {
+                                        id,
+                                        window: Arc::new(window),
+                                        exit_on_close_request,
+                                        make_visible: visible,
+                                        on_open,
+                                    },
+                                );
+                            }
+                            Control::Exit => {
+                                event_loop.exit();
+                            }
+                            Control::Crash(error) => {
+                                self.error = Some(error);
+                                event_loop.exit();
+                            }
+                        },
+                        _ => {
+                            break;
+                        }
+                    },
+                    task::Poll::Ready(_) => {
+                        event_loop.exit();
+                        break;
+                    }
+                };
+            }
+        }
+    }
+
+    #[cfg(not(target_arch = "wasm32"))]
+    {
+        let mut runner = runner;
+        let _ = event_loop.run_app(&mut runner);
+
+        runner.error.map(Err).unwrap_or(Ok(()))
+    }
+
+    #[cfg(target_arch = "wasm32")]
+    {
+        use winit::platform::web::EventLoopExtWebSys;
+        let _ = event_loop.spawn_app(runner);
+
+        Ok(())
+    }
+}
+
+#[derive(Debug)]
+enum Event<Message: 'static> {
+    WindowCreated {
+        id: window::Id,
+        window: Arc<winit::window::Window>,
+        exit_on_close_request: bool,
+        make_visible: bool,
+        on_open: oneshot::Sender<window::Id>,
+    },
+    EventLoopAwakened(winit::event::Event<Message>),
+}
+
+#[derive(Debug)]
+enum Control {
+    ChangeFlow(winit::event_loop::ControlFlow),
+    Exit,
+    Crash(Error),
+    CreateWindow {
+        id: window::Id,
+        settings: window::Settings,
+        title: String,
+        monitor: Option<winit::monitor::MonitorHandle>,
+        on_open: oneshot::Sender<window::Id>,
+    },
+}
+
+async fn run_instance<P>(
+    mut program: program::Instance<P>,
+    mut runtime: Runtime<P::Executor, Proxy<P::Message>, Action<P::Message>>,
+    mut proxy: Proxy<P::Message>,
+    mut event_receiver: mpsc::UnboundedReceiver<Event<Action<P::Message>>>,
+    mut control_sender: mpsc::UnboundedSender<Control>,
+    is_daemon: bool,
+    graphics_settings: graphics::Settings,
+    default_fonts: Vec<Cow<'static, [u8]>>,
+) where
+    P: Program + 'static,
+    P::Theme: theme::Base,
+{
+    use winit::event;
+    use winit::event_loop::ControlFlow;
+
+    let mut window_manager = WindowManager::new();
+    let mut is_window_opening = !is_daemon;
+
+    let mut compositor = None;
+    let mut events = Vec::new();
+    let mut messages = Vec::new();
+    let mut actions = 0;
+
+    let mut ui_caches = FxHashMap::default();
+    let mut user_interfaces = ManuallyDrop::new(FxHashMap::default());
+    let mut clipboard = Clipboard::unconnected();
+
+    loop {
+        // Empty the queue if possible
+        let event = if let Ok(event) = event_receiver.try_next() {
+            event
+        } else {
+            event_receiver.next().await
+        };
+
+        let Some(event) = event else {
+            break;
+        };
+
+        match event {
+            Event::WindowCreated {
+                id,
+                window,
+                exit_on_close_request,
+                make_visible,
+                on_open,
+            } => {
+                if compositor.is_none() {
+                    let (compositor_sender, compositor_receiver) =
+                        oneshot::channel();
+
+                    let create_compositor = {
+                        let window = window.clone();
+                        let mut proxy = proxy.clone();
+                        let default_fonts = default_fonts.clone();
+
+                        async move {
+                            let mut compositor =
+                                <P::Renderer as compositor::Default>::Compositor::new(graphics_settings, window).await;
+
+                            if let Ok(compositor) = &mut compositor {
+                                for font in default_fonts {
+                                    compositor.load_font(font.clone());
+                                }
+                            }
+
+                            compositor_sender
+                                .send(compositor)
+                                .ok()
+                                .expect("Send compositor");
+
+                            // HACK! Send a proxy event on completion to trigger
+                            // a runtime re-poll
+                            // TODO: Send compositor through proxy (?)
+                            {
+                                let (sender, _receiver) = oneshot::channel();
+
+                                proxy.send_action(Action::Window(
+                                    runtime::window::Action::GetLatest(sender),
+                                ));
+                            }
+                        }
+                    };
+
+                    #[cfg(target_arch = "wasm32")]
+                    wasm_bindgen_futures::spawn_local(create_compositor);
+
+                    #[cfg(not(target_arch = "wasm32"))]
+                    runtime.block_on(create_compositor);
+
+                    match compositor_receiver
+                        .await
+                        .expect("Wait for compositor")
+                    {
+                        Ok(new_compositor) => {
+                            compositor = Some(new_compositor);
+                        }
+                        Err(error) => {
+                            let _ = control_sender
+                                .start_send(Control::Crash(error.into()));
+                            continue;
+                        }
+                    }
+                }
+
+                debug::theme_changed(|| {
+                    if window_manager.is_empty() {
+                        theme::Base::palette(&program.theme(id))
+                    } else {
+                        None
+                    }
+                });
+
+                let window = window_manager.insert(
+                    id,
+                    window,
+                    &program,
+                    compositor
+                        .as_mut()
+                        .expect("Compositor must be initialized"),
+                    exit_on_close_request,
+                );
+
+                let logical_size = window.state.logical_size();
+
+                let _ = user_interfaces.insert(
+                    id,
+                    build_user_interface(
+                        &program,
+                        user_interface::Cache::default(),
+                        &mut window.renderer,
+                        logical_size,
+                        id,
+                    ),
+                );
+                let _ = ui_caches.insert(id, user_interface::Cache::default());
+
+                if make_visible {
+                    window.raw.set_visible(true);
+                }
+
+                events.push((
+                    id,
+                    core::Event::Window(window::Event::Opened {
+                        position: window.position(),
+                        size: window.size(),
+                    }),
+                ));
+
+                if clipboard.window_id().is_none() {
+                    clipboard = Clipboard::connect(window.raw.clone());
+                }
+
+                let _ = on_open.send(id);
+                is_window_opening = false;
+            }
+            Event::EventLoopAwakened(event) => {
+                match event {
+                    event::Event::NewEvents(event::StartCause::Init) => {
+                        for (_id, window) in window_manager.iter_mut() {
+                            window.raw.request_redraw();
+                        }
+                    }
+                    event::Event::NewEvents(
+                        event::StartCause::ResumeTimeReached { .. },
+                    ) => {
+                        let now = Instant::now();
+
+                        for (_id, window) in window_manager.iter_mut() {
+                            if let Some(redraw_at) = window.redraw_at {
+                                if redraw_at <= now {
+                                    window.raw.request_redraw();
+                                    window.redraw_at = None;
+                                }
+                            }
+                        }
+
+                        if let Some(redraw_at) = window_manager.redraw_at() {
+                            let _ =
+                                control_sender.start_send(Control::ChangeFlow(
+                                    ControlFlow::WaitUntil(redraw_at),
+                                ));
+                        } else {
+                            let _ = control_sender.start_send(
+                                Control::ChangeFlow(ControlFlow::Wait),
+                            );
+                        }
+                    }
+                    event::Event::PlatformSpecific(
+                        event::PlatformSpecific::MacOS(
+                            event::MacOS::ReceivedUrl(url),
+                        ),
+                    ) => {
+                        runtime.broadcast(
+                            subscription::Event::PlatformSpecific(
+                                subscription::PlatformSpecific::MacOS(
+                                    subscription::MacOS::ReceivedUrl(url),
+                                ),
+                            ),
+                        );
+                    }
+                    event::Event::UserEvent(action) => {
+                        run_action(
+                            action,
+                            &program,
+                            &mut compositor,
+                            &mut events,
+                            &mut messages,
+                            &mut clipboard,
+                            &mut control_sender,
+                            &mut user_interfaces,
+                            &mut window_manager,
+                            &mut ui_caches,
+                            &mut is_window_opening,
+                        );
+                        actions += 1;
+                    }
+                    event::Event::WindowEvent {
+                        window_id: id,
+                        event: event::WindowEvent::RedrawRequested,
+                        ..
+                    } => {
+                        let Some(compositor) = &mut compositor else {
+                            continue;
+                        };
+
+                        let Some((id, window)) =
+                            window_manager.get_mut_alias(id)
+                        else {
+                            continue;
+                        };
+
+                        let physical_size = window.state.physical_size();
+
+                        if physical_size.width == 0 || physical_size.height == 0
+                        {
+                            continue;
+                        }
+
+                        if window.viewport_version
+                            != window.state.viewport_version()
+                        {
+                            let logical_size = window.state.logical_size();
+
+                            let layout_span = debug::layout(id);
+                            let ui = user_interfaces
+                                .remove(&id)
+                                .expect("Remove user interface");
+
+                            let _ = user_interfaces.insert(
+                                id,
+                                ui.relayout(logical_size, &mut window.renderer),
+                            );
+                            layout_span.finish();
+
+                            compositor.configure_surface(
+                                &mut window.surface,
+                                physical_size.width,
+                                physical_size.height,
+                            );
+
+                            window.viewport_version =
+                                window.state.viewport_version();
+                        }
+
+                        let redraw_event = core::Event::Window(
+                            window::Event::RedrawRequested(Instant::now()),
+                        );
+
+                        let cursor = window.state.cursor();
+
+                        let ui = user_interfaces
+                            .get_mut(&id)
+                            .expect("Get user interface");
+
+                        let draw_span = debug::draw(id);
+                        let (ui_state, _) = ui.update(
+                            &[redraw_event.clone()],
+                            cursor,
+                            &mut window.renderer,
+                            &mut clipboard,
+                            &mut messages,
+                        );
+
+                        ui.draw(
+                            &mut window.renderer,
+                            window.state.theme(),
+                            &renderer::Style {
+                                text_color: window.state.text_color(),
+                            },
+                            cursor,
+                        );
+                        draw_span.finish();
+
+                        runtime.broadcast(subscription::Event::Interaction {
+                            window: id,
+                            event: redraw_event,
+                            status: core::event::Status::Ignored,
+                        });
+
+                        if let user_interface::State::Updated {
+                            redraw_request,
+                            input_method,
+                            mouse_interaction,
+                        } = ui_state
+                        {
+                            window.request_redraw(redraw_request);
+                            window.request_input_method(input_method);
+                            window.update_mouse(mouse_interaction);
+                        }
+
+                        window.draw_preedit();
+
+                        let present_span = debug::present(id);
+                        match compositor.present(
+                            &mut window.renderer,
+                            &mut window.surface,
+                            window.state.viewport(),
+                            window.state.background_color(),
+                            || window.raw.pre_present_notify(),
+                        ) {
+                            Ok(()) => {
+                                present_span.finish();
+                            }
+                            Err(error) => match error {
+                                // This is an unrecoverable error.
+                                compositor::SurfaceError::OutOfMemory => {
+                                    panic!("{:?}", error);
+                                }
+                                _ => {
+                                    present_span.finish();
+
+                                    log::error!(
+                                        "Error {error:?} when \
+                                        presenting surface."
+                                    );
+
+                                    // Try rendering all windows again next frame.
+                                    for (_id, window) in
+                                        window_manager.iter_mut()
+                                    {
+                                        window.raw.request_redraw();
+                                    }
+                                }
+                            },
+                        }
+                    }
+                    event::Event::WindowEvent {
+                        event: window_event,
+                        window_id,
+                    } => {
+                        if !is_daemon
+                            && matches!(
+                                window_event,
+                                winit::event::WindowEvent::Destroyed
+                            )
+                            && !is_window_opening
+                            && window_manager.is_empty()
+                        {
+                            control_sender
+                                .start_send(Control::Exit)
+                                .expect("Send control action");
+
+                            continue;
+                        }
+
+                        let Some((id, window)) =
+                            window_manager.get_mut_alias(window_id)
+                        else {
+                            continue;
+                        };
+
+                        if matches!(
+                            window_event,
+                            winit::event::WindowEvent::Resized(_)
+                        ) {
+                            window.raw.request_redraw();
+                        }
+
+                        if matches!(
+                            window_event,
+                            winit::event::WindowEvent::CloseRequested
+                        ) && window.exit_on_close_request
+                        {
+                            run_action(
+                                Action::Window(runtime::window::Action::Close(
+                                    id,
+                                )),
+                                &program,
+                                &mut compositor,
+                                &mut events,
+                                &mut messages,
+                                &mut clipboard,
+                                &mut control_sender,
+                                &mut user_interfaces,
+                                &mut window_manager,
+                                &mut ui_caches,
+                                &mut is_window_opening,
+                            );
+                        } else {
+                            window.state.update(&window.raw, &window_event);
+
+                            if let Some(event) = conversion::window_event(
+                                window_event,
+                                window.state.scale_factor(),
+                                window.state.modifiers(),
+                            ) {
+                                events.push((id, event));
+                            }
+                        }
+                    }
+                    event::Event::AboutToWait => {
+                        if actions > 0 {
+                            proxy.free_slots(actions);
+                            actions = 0;
+                        }
+
+                        if events.is_empty()
+                            && messages.is_empty()
+                            && window_manager.is_idle()
+                        {
+                            continue;
+                        }
+
+                        let mut uis_stale = false;
+
+                        for (id, window) in window_manager.iter_mut() {
+                            let interact_span = debug::interact(id);
+                            let mut window_events = vec![];
+
+                            events.retain(|(window_id, event)| {
+                                if *window_id == id {
+                                    window_events.push(event.clone());
+                                    false
+                                } else {
+                                    true
+                                }
+                            });
+
+                            if window_events.is_empty() && messages.is_empty() {
+                                continue;
+                            }
+
+                            let (ui_state, statuses) = user_interfaces
+                                .get_mut(&id)
+                                .expect("Get user interface")
+                                .update(
+                                    &window_events,
+                                    window.state.cursor(),
+                                    &mut window.renderer,
+                                    &mut clipboard,
+                                    &mut messages,
+                                );
+
+                            #[cfg(feature = "unconditional-rendering")]
+                            window.request_redraw(
+                                window::RedrawRequest::NextFrame,
+                            );
+
+                            match ui_state {
+                                user_interface::State::Updated {
+                                    redraw_request: _redraw_request,
+                                    mouse_interaction,
+                                    ..
+                                } => {
+                                    window.update_mouse(mouse_interaction);
+
+                                    #[cfg(not(
+                                        feature = "unconditional-rendering"
+                                    ))]
+                                    window.request_redraw(_redraw_request);
+                                }
+                                user_interface::State::Outdated => {
+                                    uis_stale = true;
+                                }
+                            }
+
+                            for (event, status) in window_events
+                                .into_iter()
+                                .zip(statuses.into_iter())
+                            {
+                                runtime.broadcast(
+                                    subscription::Event::Interaction {
+                                        window: id,
+                                        event,
+                                        status,
+                                    },
+                                );
+                            }
+
+                            interact_span.finish();
+                        }
+
+                        for (id, event) in events.drain(..) {
+                            runtime.broadcast(
+                                subscription::Event::Interaction {
+                                    window: id,
+                                    event,
+                                    status: core::event::Status::Ignored,
+                                },
+                            );
+                        }
+
+                        if !messages.is_empty() || uis_stale {
+                            let cached_interfaces: FxHashMap<
+                                window::Id,
+                                user_interface::Cache,
+                            > = ManuallyDrop::into_inner(user_interfaces)
+                                .drain()
+                                .map(|(id, ui)| (id, ui.into_cache()))
+                                .collect();
+
+                            update(&mut program, &mut runtime, &mut messages);
+
+                            for (id, window) in window_manager.iter_mut() {
+                                window.state.synchronize(
+                                    &program,
+                                    id,
+                                    &window.raw,
+                                );
+
+                                window.raw.request_redraw();
+                            }
+
+                            debug::theme_changed(|| {
+                                window_manager.first().and_then(|window| {
+                                    theme::Base::palette(window.state.theme())
+                                })
+                            });
+
+                            user_interfaces =
+                                ManuallyDrop::new(build_user_interfaces(
+                                    &program,
+                                    &mut window_manager,
+                                    cached_interfaces,
+                                ));
+                        }
+
+                        if let Some(redraw_at) = window_manager.redraw_at() {
+                            let _ =
+                                control_sender.start_send(Control::ChangeFlow(
+                                    ControlFlow::WaitUntil(redraw_at),
+                                ));
+                        } else {
+                            let _ = control_sender.start_send(
+                                Control::ChangeFlow(ControlFlow::Wait),
+                            );
+                        }
+                    }
+                    _ => {}
+                }
+            }
+        }
+    }
+
+    let _ = ManuallyDrop::into_inner(user_interfaces);
+}
+
+/// Builds a window's [`UserInterface`] for the [`Program`].
+fn build_user_interface<'a, P: Program>(
+    program: &'a program::Instance<P>,
+    cache: user_interface::Cache,
+    renderer: &mut P::Renderer,
+    size: Size,
+    id: window::Id,
+) -> UserInterface<'a, P::Message, P::Theme, P::Renderer>
+where
+    P::Theme: theme::Base,
+{
+    let view_span = debug::view(id);
+    let view = program.view(id);
+    view_span.finish();
+
+    let layout_span = debug::layout(id);
+    let user_interface = UserInterface::build(view, size, cache, renderer);
+    layout_span.finish();
+
+    user_interface
+}
+
+fn update<P: Program, E: Executor>(
+    program: &mut program::Instance<P>,
+    runtime: &mut Runtime<E, Proxy<P::Message>, Action<P::Message>>,
+    messages: &mut Vec<P::Message>,
+) where
+    P::Theme: theme::Base,
+{
+    for message in messages.drain(..) {
+        let task = runtime.enter(|| program.update(message));
+
+        if let Some(stream) = runtime::task::into_stream(task) {
+            runtime.run(stream);
+        }
+    }
+
+    let subscription = runtime.enter(|| program.subscription());
+    let recipes = subscription::into_recipes(subscription.map(Action::Output));
+
+    runtime.track(recipes);
+}
+
+fn run_action<P, C>(
+    action: Action<P::Message>,
+    program: &program::Instance<P>,
+    compositor: &mut Option<C>,
+    events: &mut Vec<(window::Id, core::Event)>,
+    messages: &mut Vec<P::Message>,
+    clipboard: &mut Clipboard,
+    control_sender: &mut mpsc::UnboundedSender<Control>,
+    interfaces: &mut FxHashMap<
+        window::Id,
+        UserInterface<'_, P::Message, P::Theme, P::Renderer>,
+    >,
+    window_manager: &mut WindowManager<P, C>,
+    ui_caches: &mut FxHashMap<window::Id, user_interface::Cache>,
+    is_window_opening: &mut bool,
+) where
+    P: Program,
+    C: Compositor<Renderer = P::Renderer> + 'static,
+    P::Theme: theme::Base,
+{
+    use crate::runtime::clipboard;
+    use crate::runtime::system;
+    use crate::runtime::window;
+
+    match action {
+        Action::Output(message) => {
+            messages.push(message);
+        }
+        Action::Clipboard(action) => match action {
+            clipboard::Action::Read { target, channel } => {
+                let _ = channel.send(clipboard.read(target));
+            }
+            clipboard::Action::Write { target, contents } => {
+                clipboard.write(target, contents);
+            }
+        },
+        Action::Window(action) => match action {
+            window::Action::Open(id, settings, channel) => {
+                let monitor = window_manager.last_monitor();
+
+                control_sender
+                    .start_send(Control::CreateWindow {
+                        id,
+                        settings,
+                        title: program.title(id),
+                        monitor,
+                        on_open: channel,
+                    })
+                    .expect("Send control action");
+
+                *is_window_opening = true;
+            }
+            window::Action::Close(id) => {
+                let _ = ui_caches.remove(&id);
+                let _ = interfaces.remove(&id);
+
+                if let Some(window) = window_manager.remove(id) {
+                    if clipboard.window_id() == Some(window.raw.id()) {
+                        *clipboard = window_manager
+                            .first()
+                            .map(|window| window.raw.clone())
+                            .map(Clipboard::connect)
+                            .unwrap_or_else(Clipboard::unconnected);
+                    }
+
+                    events.push((
+                        id,
+                        core::Event::Window(core::window::Event::Closed),
+                    ));
+                }
+
+                if window_manager.is_empty() {
+                    *compositor = None;
+                }
+            }
+            window::Action::GetOldest(channel) => {
+                let id =
+                    window_manager.iter_mut().next().map(|(id, _window)| id);
+
+                let _ = channel.send(id);
+            }
+            window::Action::GetLatest(channel) => {
+                let id =
+                    window_manager.iter_mut().last().map(|(id, _window)| id);
+
+                let _ = channel.send(id);
+            }
+            window::Action::Drag(id) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let _ = window.raw.drag_window();
+                }
+            }
+            window::Action::DragResize(id, direction) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let _ = window.raw.drag_resize_window(
+                        conversion::resize_direction(direction),
+                    );
+                }
+            }
+            window::Action::Resize(id, size) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let _ = window.raw.request_inner_size(
+                        winit::dpi::LogicalSize {
+                            width: size.width,
+                            height: size.height,
+                        },
+                    );
+                }
+            }
+            window::Action::SetMinSize(id, size) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_min_inner_size(size.map(|size| {
+                        winit::dpi::LogicalSize {
+                            width: size.width,
+                            height: size.height,
+                        }
+                    }));
+                }
+            }
+            window::Action::SetMaxSize(id, size) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_max_inner_size(size.map(|size| {
+                        winit::dpi::LogicalSize {
+                            width: size.width,
+                            height: size.height,
+                        }
+                    }));
+                }
+            }
+            window::Action::SetResizeIncrements(id, increments) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_resize_increments(increments.map(|size| {
+                        winit::dpi::LogicalSize {
+                            width: size.width,
+                            height: size.height,
+                        }
+                    }));
+                }
+            }
+            window::Action::SetResizable(id, resizable) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_resizable(resizable);
+                }
+            }
+            window::Action::GetSize(id, channel) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let size = window
+                        .raw
+                        .inner_size()
+                        .to_logical(window.raw.scale_factor());
+
+                    let _ = channel.send(Size::new(size.width, size.height));
+                }
+            }
+            window::Action::GetMaximized(id, channel) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let _ = channel.send(window.raw.is_maximized());
+                }
+            }
+            window::Action::Maximize(id, maximized) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_maximized(maximized);
+                }
+            }
+            window::Action::GetMinimized(id, channel) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let _ = channel.send(window.raw.is_minimized());
+                }
+            }
+            window::Action::Minimize(id, minimized) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_minimized(minimized);
+                }
+            }
+            window::Action::GetPosition(id, channel) => {
+                if let Some(window) = window_manager.get(id) {
+                    let position = window
+                        .raw
+                        .outer_position()
+                        .map(|position| {
+                            let position = position
+                                .to_logical::<f32>(window.raw.scale_factor());
+
+                            Point::new(position.x, position.y)
+                        })
+                        .ok();
+
+                    let _ = channel.send(position);
+                }
+            }
+            window::Action::GetScaleFactor(id, channel) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let scale_factor = window.raw.scale_factor();
+
+                    let _ = channel.send(scale_factor as f32);
+                }
+            }
+            window::Action::Move(id, position) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_outer_position(
+                        winit::dpi::LogicalPosition {
+                            x: position.x,
+                            y: position.y,
+                        },
+                    );
+                }
+            }
+            window::Action::SetMode(id, mode) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_visible(conversion::visible(mode));
+                    window.raw.set_fullscreen(conversion::fullscreen(
+                        window.raw.current_monitor(),
+                        mode,
+                    ));
+                }
+            }
+            window::Action::SetIcon(id, icon) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_window_icon(conversion::icon(icon));
+                }
+            }
+            window::Action::GetMode(id, channel) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let mode = if window.raw.is_visible().unwrap_or(true) {
+                        conversion::mode(window.raw.fullscreen())
+                    } else {
+                        core::window::Mode::Hidden
+                    };
+
+                    let _ = channel.send(mode);
+                }
+            }
+            window::Action::ToggleMaximize(id) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_maximized(!window.raw.is_maximized());
+                }
+            }
+            window::Action::ToggleDecorations(id) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.set_decorations(!window.raw.is_decorated());
+                }
+            }
+            window::Action::RequestUserAttention(id, attention_type) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.request_user_attention(
+                        attention_type.map(conversion::user_attention),
+                    );
+                }
+            }
+            window::Action::GainFocus(id) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window.raw.focus_window();
+                }
+            }
+            window::Action::SetLevel(id, level) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    window
+                        .raw
+                        .set_window_level(conversion::window_level(level));
+                }
+            }
+            window::Action::ShowSystemMenu(id) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    if let mouse::Cursor::Available(point) =
+                        window.state.cursor()
+                    {
+                        window.raw.show_window_menu(
+                            winit::dpi::LogicalPosition {
+                                x: point.x,
+                                y: point.y,
+                            },
+                        );
+                    }
+                }
+            }
+            window::Action::GetRawId(id, channel) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let _ = channel.send(window.raw.id().into());
+                }
+            }
+            window::Action::RunWithHandle(id, f) => {
+                use window::raw_window_handle::HasWindowHandle;
+
+                if let Some(handle) = window_manager
+                    .get_mut(id)
+                    .and_then(|window| window.raw.window_handle().ok())
+                {
+                    f(handle);
+                }
+            }
+            window::Action::Screenshot(id, channel) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    if let Some(compositor) = compositor {
+                        let bytes = compositor.screenshot(
+                            &mut window.renderer,
+                            window.state.viewport(),
+                            window.state.background_color(),
+                        );
+
+                        let _ = channel.send(core::window::Screenshot::new(
+                            bytes,
+                            window.state.physical_size(),
+                            window.state.viewport().scale_factor(),
+                        ));
+                    }
+                }
+            }
+            window::Action::EnableMousePassthrough(id) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let _ = window.raw.set_cursor_hittest(false);
+                }
+            }
+            window::Action::DisableMousePassthrough(id) => {
+                if let Some(window) = window_manager.get_mut(id) {
+                    let _ = window.raw.set_cursor_hittest(true);
+                }
+            }
+        },
+        Action::System(action) => match action {
+            system::Action::QueryInformation(_channel) => {
+                #[cfg(feature = "system")]
+                {
+                    if let Some(compositor) = compositor {
+                        let graphics_info = compositor.fetch_information();
+
+                        let _ = std::thread::spawn(move || {
+                            let information =
+                                crate::system::information(graphics_info);
+
+                            let _ = _channel.send(information);
+                        });
+                    }
+                }
+            }
+        },
+        Action::Widget(operation) => {
+            let mut current_operation = Some(operation);
+
+            while let Some(mut operation) = current_operation.take() {
+                for (id, ui) in interfaces.iter_mut() {
+                    if let Some(window) = window_manager.get_mut(*id) {
+                        ui.operate(&window.renderer, operation.as_mut());
+                    }
+                }
+
+                match operation.finish() {
+                    operation::Outcome::None => {}
+                    operation::Outcome::Some(()) => {}
+                    operation::Outcome::Chain(next) => {
+                        current_operation = Some(next);
+                    }
+                }
+            }
+        }
+        Action::LoadFont { bytes, channel } => {
+            if let Some(compositor) = compositor {
+                // TODO: Error handling (?)
+                compositor.load_font(bytes.clone());
+
+                let _ = channel.send(Ok(()));
+            }
+        }
+        Action::Exit => {
+            control_sender
+                .start_send(Control::Exit)
+                .expect("Send control action");
+        }
+    }
+}
+
+/// Build the user interface for every window.
+pub fn build_user_interfaces<'a, P: Program, C>(
+    program: &'a program::Instance<P>,
+    window_manager: &mut WindowManager<P, C>,
+    mut cached_user_interfaces: FxHashMap<window::Id, user_interface::Cache>,
+) -> FxHashMap<window::Id, UserInterface<'a, P::Message, P::Theme, P::Renderer>>
+where
+    C: Compositor<Renderer = P::Renderer>,
+    P::Theme: theme::Base,
+{
+    cached_user_interfaces
+        .drain()
+        .filter_map(|(id, cache)| {
+            let window = window_manager.get_mut(id)?;
+
+            Some((
+                id,
+                build_user_interface(
+                    program,
+                    cache,
+                    &mut window.renderer,
+                    window.state.logical_size(),
+                    id,
+                ),
+            ))
+        })
+        .collect()
+}
+
+/// Returns true if the provided event should cause a [`Program`] to
+/// exit.
+pub fn user_force_quit(
+    event: &winit::event::WindowEvent,
+    _modifiers: winit::keyboard::ModifiersState,
+) -> bool {
+    match event {
+        #[cfg(target_os = "macos")]
+        winit::event::WindowEvent::KeyboardInput {
+            event:
+                winit::event::KeyEvent {
+                    logical_key: winit::keyboard::Key::Character(c),
+                    state: winit::event::ElementState::Pressed,
+                    ..
+                },
+            ..
+        } if c == "q" && _modifiers.super_key() => true,
+        _ => false,
+    }
+}
